### 2023-06-08 #115 @nwlandry
XGI
Submitting Author: Name (@nwlandry)
All current maintainers: (@leotrs, @maximelucas, @iaciac, @lordgrilo, @acuschwarze, @thomasrobiglio, and @alpatania)
Package Name: XGI
One-Line Description of Package: XGI is a Python package for higher-order networks.
Repository Link: https://github.com/xgi-org/xgi
Version submitted: 0.7
Editor: Szymon Moliński (@SimonMolinsky)
Reviewer 1: Nhat (Jonny) Tran (@JonnyTran)
Reviewer 2: Marta Leszczyńska (@Reckony)
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8365068.svg)](https://doi.org/10.5281/zenodo.8365068)
Version accepted: 0.7.4

---

- Include a brief paragraph describing what your package does: Comple**X** **G**roup **I**nteractions (XGI) is a library for analyzing higher-order networks. Such networks are used to model interactions of arbitrary size between entities in a complex system. This library provides methods for building hypergraphs and simplicial complexes; algorithms to analyze their structure, visualize them, and simulate dynamical processes on them; and a collection of higher-order datasets. XGI is implemented in pure Python and integrates with the rest of the Python scientific stack. XGI is designed and developed by network scientists with the needs of network scientists in mind.

- [x] NumFOCUS
    XGI is a [NumFOCUS Affiliated Project](https://numfocus.org/sponsored-projects/affiliated-projects).

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

 The target audience is network scientists across multiple disciplines, particularly those interested in networks that contain higher-order interactions. The scientific application of this package is provide a common language for network scientists to work on projects related to higher-order interactions by providing common data structures, algorithms, and basic visualization tools as well as standardized datasets.

There are several existing packages to represent and analyze higher-order networks: `HyperNetX`, `Hypergraphx`, `Hypergraph Analysis Toolbox (HAT)`, and `Reticula` in Python, `SimpleHypergraphs.jl` and `HyperGraphs.jl` in Julia, and `hyperG` in R. XGI is a valuable addition to the current suite of library available. First, XGI is implemented in pure Python, ensuring interoperability and easy installation across operating systems. Second, like several of the packages listed, XGI has a well-documented codebase and tutorials designed to make the learning process intuitive. Third, in contrast to existing packages, XGI contains a `stats` module enabling researchers to easily access established nodal and edge quantities, and even define custom quantities. Fourth, XGI offers data structures for hypergraphs, directed hypergraphs, and simplicial complexes, which allows users to explore a wider range of interaction models than comparable packages. Lastly, XGI integrates higher-order datasets with its interface, providing a standard format in which to store hypergraphs with attributes and a data repository with corresponding functions to load these datasets.
