### 2023-06-09 fe00623
bug fix for draw_dihypergraph
fix: ability to handle integer nodes and isolated nodes
test: test for DH with str node labels and isolated nodes
style: run black
