+++ docs/source/api/classes/xgi.classes.dihypergraph.DiHypergraph.rst
@@ -3,6 +3,9 @@
 
 .. currentmodule:: xgi.classes.dihypergraph
 
+.. warning::
+    This is an experimental module. It has been throughly tested in development
+    but has not been used in research. It is not yet stable and may change in the future.
 .. autoclass:: DiHypergraph
    :show-inheritance:
    :members:
