### 2023-06-09 8c8728e
fix: separated test for str and isolated nodes
