### 2023-06-09 81efa70
remove mid sentence
