### 2023-06-09 #115 @nwlandry
Hi @NickleDave - yes, it was an intentional submission, but I didn't realize all the missing fields. Sorry about that! I will correct this now. This is already published in JOSS, so I won't fill out the JOSS related items. Thanks so much.
