### 2023-06-12 00df310
feat: ability to create SC from simplex dict
