### 2023-06-12 82fa31b
added from_simplex_dict to __all__
