### 2023-06-12 a1ef759
docs: updated docs about incoming data in HG and SC classes
