+++ xgi/classes/simplicialcomplex.py
@@ -45,7 +45,7 @@ class SimplicialComplex(Hypergraph):
         * simplex dictionary
         * 2-column Pandas dataframe (bipartite edges)
         * Scipy/Numpy incidence matrix
-        * SimplicialComplex object.
+        * SimplicialComplex object
         * Hypergraph object
 
     **attr : dict, optional
