+++ docs/source/api/generators/xgi.generators.randomizing.rst
@@ -1,0 +1,13 @@
+xgi.generators.random
+=====================
+
+.. currentmodule:: xgi.generators.random
+
+.. automodule:: xgi.generators.random
+   
+    .. rubric:: Functions
+   
+    .. autofunction:: chung_lu_hypergraph
+    .. autofunction:: dcsbm_hypergraph
+    .. autofunction:: random_hypergraph
+    .. autofunction:: watts_strogatz_hypergraph
