+++ xgi/generators/__init__.py
@@ -1,4 +1,12 @@
-from . import classic, uniform, random, simple, lattice, simplicial_complexes, randomizing
+from . import (
+    classic,
+    uniform,
+    random,
+    simple,
+    lattice,
+    simplicial_complexes,
+    randomizing,
+)
 from .classic import *
 from .random import *
 from .uniform import *
