### 2023-06-13 bf50618
fix: HG using SC as data
