+++ xgi/generators/randomizing.py
@@ -6,11 +6,13 @@ hypergraph).
 """
 
 import random
+from copy import deepcopy
 
 import xgi
 
 __all__ = [
     "shuffle_hyperedges",
+    "node_swap",
 ]
 
 
@@ -31,9 +33,14 @@ def shuffle_hyperedges(S, order, p):
     H: xgi.HyperGraph
         Hypergraph with edges of order d shuffled
 
+    Note
+    ----
+    By shuffling hyperedges in a simplicial complex, it will in general lose
+    its "simpliciality" and become a hypergraph.
+
     Reference
     ---------
-    Zhang, Y., Lucas, M. and Battiston, F., 2023.
+    Zhang, Y.*, Lucas, M.* and Battiston, F., 2023.
     "Higher-order interactions shape collective dynamics differently
     in hypergraphs and simplicial complexes."
     Nature Communications, 14(1), p.1605.
@@ -41,7 +48,8 @@ def shuffle_hyperedges(S, order, p):
 
     Example
     -------
-
+    >>> S = xgi.random_simplicial_complex(50, [0.1, 0.01, 0.001], seed=1)
+    >>> H = xgi.shuffle_hyperedges(S, order=2, p=0.5)
 
     """
 
@@ -69,8 +77,92 @@ def shuffle_hyperedges(S, order, p):
                 new_hyperedge = tuple(random.sample(nodes, order + 1))
             H.add_edge(new_hyperedge)
 
-    assert H.num_nodes == S.num_nodes
-    assert xgi.num_edges_order(H, 1) == xgi.num_edges_order(S, 1)
-    assert xgi.num_edges_order(H, 2) == xgi.num_edges_order(S, 2)
-
     return H
+
+
+def node_swap(H, nid1, nid2, id_temp=-1, order=None):
+    """Swap node `nid1` and node `nid2` in all edges of order `order`.
+
+    Parameters
+    ----------
+    H: HyperGraph
+        Hypergraph to consider
+    nid1: node ID
+        ID of first node to swap
+    nid2: node ID
+        ID of second node to swap
+    id_temp: node ID
+        Temporary ID given to nodes when swapping
+    order: {int, None}, default: None
+        If None, consider all orders. If an integer,
+        consider edges of that order.
+
+    Returns
+    -------
+    HH: HyperGraph
+
+    Reference
+    ---------
+    Zhang, Y.*, Lucas, M.* and Battiston, F., 2023.
+    "Higher-order interactions shape collective dynamics differently
+    in hypergraphs and simplicial complexes."
+    Nature Communications, 14(1), p.1605.
+    https://doi.org/10.1038/s41467-023-37190-9
+
+    """
+
+    # check that node ids are in hypergraph
+    if not nid1 in H: 
+        raise ValueError(f"Node {nid1} is not in hypergraph H")
+    if not nid2 in H: 
+        raise ValueError(f"Node {nid2} is not in hypergraph H")
+
+    if order is not None:
+        if (order + 1) not in xgi.unique_edge_sizes(H):
+            raise ValueError(f"There is no hyperedge of order {order} is this hypergraph.")
+
+    # make sure id_temps does not exist yet
+    while id_temp in H.edges:
+        id_temp -= 1
+
+    # get edges of given order
+    if order:
+        edge_dict = H.edges.filterby("order", order).members(dtype=dict).copy()
+    else: # includes order 0
+        edge_dict = H.edges.members(dtype=dict).copy()
+
+    # check that node ids exist in those edges
+    if H.nodes.degree(order=order)[nid1] == 0:
+        raise ValueError(f"Node {nid1} is not part of any hyperedge of the specified order")
+    if H.nodes.degree(order=order)[nid2] == 0:
[…]
