### 2023-06-13 4d7a2f8
test: add test for HG constructor with SC input data
