### 2023-06-13 4b7aadf
test: test for SC constructor with H input data + minor fix
