### 2023-06-13 #115 @SimonMolinsky
Hi @nwlandry 👋

I'll go through the submission process with you. Thanks for filling out the survey! Now it's my turn ♟️ I will perform initial checks of the package, and if everything is fine, the review process begins. Please, ping me in cases related to the package or review process.
