### 2023-06-13 #115 @SimonMolinsky
- [ ] The package imports properly into a standard Python environment `import package-name`.
  - [ ] **Code of Conduct** The package has a `Code of Conduct` file.

---
---
