+++ CONTRIBUTORS.md
@@ -1,0 +1,23 @@
+The XGI project has been helped by invaluable contributions from many members of the community.
+
+# Core team
+
+@nwlandry
+@maximelucas
+@iaciac
+@lordgrilo
+@acuschwarze
+@alpatania
+@leotrs
+@thomasrobiglio
+
+# Core team alumni
+@mcontisc
+
+# Contributors
+@acombretrenouard
+@tlarock
+@marconurisso
+@saad1282
+@arnaudon
+@goznalo-git
