### 2023-06-13 468b21e
remove contributors from every file
