+++ CONTRIBUTORS.md
@@ -1,23 +1,25 @@
 The XGI project has been helped by invaluable contributions from many members of the community.
 
-# Core team
+## Core team
 
-@nwlandry
-@maximelucas
-@iaciac
-@lordgrilo
-@acuschwarze
-@alpatania
-@leotrs
-@thomasrobiglio
+* [@nwlandry](https://github.com/nwlandry)
+* [@maximelucas](https://github.com/maximelucas)
+* [@iaciac](https://github.com/iaciac)
+* [@lordgrilo](https://github.com/lordgrilo)
+* [@acuschwarze](https://github.com/acuschwarze)
+* [@alpatania](https://github.com/alpatania)
+* [@leotrs](https://github.com/leotrs)
+* [@thomasrobiglio](https://github.com/thomasrobiglio)
 
-# Core team alumni
-@mcontisc
+## Core team alumni
 
-# Contributors
-@acombretrenouard
-@tlarock
-@marconurisso
-@saad1282
-@arnaudon
-@goznalo-git
+* [@mcontisc](https://github.com/mcontisc)
+
+## Contributors
+
+* [@acombretrenouard](https://github.com/acombretrenouard)
+* [@tlarock](https://github.com/tlarock)
+* [@marconurisso](https://github.com/marconurisso)
+* [@saad1282](https://github.com/saad1282)
+* [@arnaudon](https://github.com/arnaudon)
+* [@goznalo-git](https://github.com/goznalo-git)
