### 2023-06-13 4667c7d
simplify img tag
