### 2023-06-13 96cb3c5
Convert from simplex dict to SC (#397)
* feat: ability to create SC from simplex dict

* test: modified test to include dict -> SC

* style: run black

* added from_simplex_dict to __all__

* docs: minor change

* docs: corrected grammar mistake

* docs: updated docs about incoming data in HG and SC classes

* docs: minor change
