+++ tests/classes/test_simplicialcomplex.py
@@ -10,18 +10,18 @@ def test_constructor(edgelist5, dict5, incidence5, dataframe5):
     S_list = xgi.SimplicialComplex(edgelist5)
     S_df = xgi.SimplicialComplex(dataframe5)
     S_sc = xgi.SimplicialComplex(S_list)
+    S_dict = xgi.SimplicialComplex(dict5)
 
-    with pytest.raises(XGIError):
-        _ = xgi.SimplicialComplex(dict5)
     with pytest.raises(XGIError):
         _ = xgi.SimplicialComplex(incidence5)
 
-    assert set(S_list.nodes) == set(S_df.nodes) == set(S_sc.nodes)
-    assert set(S_list.edges) == set(S_df.edges) == set(S_sc.edges)
+    assert set(S_list.nodes) == set(S_df.nodes) == set(S_sc.nodes) == set(S_dict.nodes)
+    assert set(S_list.edges) == set(S_df.edges) == set(S_sc.edges) == set(S_dict.edges)
     assert (
         set(S_list.edges.members(0))
         == set(S_df.edges.members(0))
         == set(S_sc.edges.members(0))
+        == set(S_dict.edges.members(0))
     )
 
     with pytest.raises(XGIError):
