+++ xgi/classes/simplicialcomplex.py
@@ -45,7 +45,8 @@ class SimplicialComplex(Hypergraph):
         * simplex dictionary
         * 2-column Pandas dataframe (bipartite edges)
         * Scipy/Numpy incidence matrix
-        * SimplicialComplex object.
+        * SimplicialComplex object
+        * Hypergraph object
 
     **attr : dict, optional
         Attributes to add to the simplicial complex as key, value pairs.
