### 2023-06-13 6418edc
Update docs (#401)
* moved contributors to a separate file.

* Rename CONTRIBUTING

* remove contributors from every file

* update docs

* added links

* added pictures, sorted alphabetically

* simplify img tag

* style: format with black
