+++ CONTRIBUTORS.md
@@ -1,0 +1,40 @@
+The XGI project has been helped by invaluable contributions from many members of the community. All listings are alphabetical.
+
+## Core team
+
+<img src="https://github.com/acuschwarze.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@acuschwarze](https://github.com/acuschwarze)
+
+<img src="https://github.com/alpatania.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@alpatania](https://github.com/alpatania)
+
+<img src="https://github.com/iaciac.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@iaciac](https://github.com/iaciac)
+
+<img src="https://github.com/leotrs.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@leotrs](https://github.com/leotrs)
+
+<img src="https://github.com/lordgrilo.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@lordgrilo](https://github.com/lordgrilo)
+
+<img src="https://github.com/maximelucas.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@maximelucas](https://github.com/maximelucas)
+
+<img src="https://github.com/nwlandry.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@nwlandry](https://github.com/nwlandry)
+
+
+<img src="https://github.com/thomasrobiglio.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@thomasrobiglio](https://github.com/thomasrobiglio)
+
+
+## Core team alumni
+
+<img src="https://github.com/mcontisc.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@mcontisc](https://github.com/mcontisc)
+
+
+## Contributors
+
+<img src="https://github.com/acombretrenouard.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/>  [@acombretrenouard](https://github.com/acombretrenouard)
+
+<img src="https://github.com/arnaudon.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@arnaudon](https://github.com/arnaudon)
+
+<img src="https://github.com/goznalo-git.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/>  [@goznalo-git](https://github.com/goznalo-git)
+
+<img src="https://github.com/marconurisso.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/>  [@marconurisso](https://github.com/marconurisso)
+
+<img src="https://github.com/saad1282.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@saad1282](https://github.com/saad1282)
+
+<img src="https://github.com/tlarock.png" width="30" height="30" style="border-radius: 50%; vertical-align: middle;"/> [@tlarock](https://github.com/tlarock)
