+++ MANIFEST.in
@@ -1,6 +1,7 @@
 include MANIFEST.in
 include setup.py
-include CONTRIBUTING.md
+include HOW_TO_CONTRIBUTE.md
+include CONTRIBUTORS.md
 include long_description.rst
 include README.md
 include LICENSE.md
