### 2023-06-13 dfff361
consistency in convert
*make convert_to_hypergraph consistent between constructor and standalone

*point to from_max_simplices in the docs

*adjust test
