+++ xgi/convert.py
@@ -69,6 +69,10 @@ def convert_to_hypergraph(data, create_using=None):
     Hypergraph object
         A hypergraph constructed from the data
 
+    See Also
+    --------
+    from_max_simplices : Constructs a hypergraph from the maximal simplices of a simplicial complex.
+
     """
     if data is None:
         return empty_hypergraph(create_using)
@@ -89,9 +93,6 @@ def convert_to_hypergraph(data, create_using=None):
         H._hypergraph = deepcopy(data._hypergraph)
         if not isinstance(create_using, DiHypergraph):
             return H
-
-    elif isinstance(data, SimplicialComplex) and create_using is None:
-        return from_max_simplices(data)
 
     elif isinstance(data, SimplicialComplex):
         H = empty_hypergraph(create_using)
