### 2023-06-13 00e14ec
Move the contributors file to readthedocs
