+++ docs/source/contributors.rst
@@ -1,0 +1,125 @@
+Contributors
+============
+
+The XGI project has been helped by invaluable contributions from many members of the community. All listings are alphabetical.
+
+Core team
+---------
+
+.. raw:: html
+
+    <p>
+        <img src="https://github.com/acuschwarze.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/acuschwarze">@acuschwarze</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/alpatania.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/alpatania">@alpatania</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/iaciac.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/iaciac">@iaciac</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/leotrs.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/leotrs">@leotrs</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/lordgrilo.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/lordgrilo">@lordgrilo</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/maximelucas.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/maximelucas">@maximelucas</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/nwlandry.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/nwlandry">@nwlandry</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/thomasrobiglio.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/thomasrobiglio">@thomasrobiglio</a>
+    </p>
+
+
+Core team alumni
+----------------
+
+.. raw:: html
+
+    <p>
+        <img src="https://github.com/mcontisc.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/mcontisc">@mcontisc</a>
+    </p>
+
+Contributors
+------------
+
+.. raw:: html
+
+    <p>
+        <img src="https://github.com/acombretrenouard.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/acombretrenouard">@acombretrenouard</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/arnaudon.png"
+                width="50" height="50" hspace="5"
+                style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/arnaudon">@arnaudon</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/goznalo-git.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/goznalo-git">@goznalo-git</a>
+    </p>
+
+    <p>
+        <img src="https://github.com/marconurisso.png"
[…]
