+++ README.md
@@ -9,8 +9,8 @@ Comple**X** **G**roup **I**nteractions (**XGI**) is a Python package for the rep
 * [**Source**](../../)
 * [**Bug reports**](../../issues)
 * [**GitHub Discussions**](../../discussions)
-* [**Documentation**](https://xgi.readthedocs.io/en/latest/)
-* [**Contributors**](CONTRIBUTORS.md)
+* [**Documentation**](https://xgi.readthedocs.io)
+* [**Contributors**](https://xgi.readthedocs.io/en/stable/contributors.html)
 
 Sign up for our [mailing list](http://eepurl.com/igE6ez) and follow XGI on [Twitter](https://twitter.com/xginets) or [Mastodon](https://mathstodon.xyz/@xginets)!
 
@@ -93,9 +93,11 @@ The XGI package has been supported by NSF Grant 2121905, ["HNDS-I: Using Hypergr
 
 ## Other Resources
 This library may not meet your needs and if this is this case, consider checking out these other resources:
-* [HyperNetX](https://pnnl.github.io/HyperNetX): A package in Python for representing, analyzing, and visualizing hypergraphs.
-* [Reticula](https://docs.reticula.network/): A package with a Python wrapper of C++ functions for representing, analyzing, and visualizing temporal and static graphs and hypergraphs.
-* [SimpleHypergraphs.jl](https://pszufe.github.io/SimpleHypergraphs.jl/v0.1/): A package in Julia for representing, analyzing, and generating hypergraphs.
-* [HyperGraphs.jl](https://github.com/lpmdiaz/HyperGraphs.jl): A package in Julia for representing, analyzing, and generating hypergraphs which may be oriented and weighted.
-* [hyperG](https://cran.r-project.org/package=HyperG): A package in R for storing and analyzing hypergraphs
-* [NetworkX](https://networkx.org/): A package in Python for representing, analyzing, and visualizing networks.
+* [HyperNetX](https://pnnl.github.io/HyperNetX): A Python package for representing, analyzing, and visualizing hypergraphs.
+* [Hypergraph Analysis Toolbox (HAT)](https://hypergraph-analysis-toolbox.readthedocs.io): A Python/Matlab package for hypergraph construction, visualization, and analysis (Especially for Pore-C data).
+* [halp](http://murali-group.github.io/halp): A Python package with directed and undirected hypergraph implementations and several algorithms.
+* [Reticula](https://docs.reticula.network): A Python package wrapping C++ functions for representing, analyzing, and visualizing temporal and static graphs and hypergraphs.
+* [SimpleHypergraphs.jl](https://pszufe.github.io/SimpleHypergraphs.jl/v0.1): A Julia package for representing, analyzing, and generating hypergraphs.
+* [HyperGraphs.jl](https://github.com/lpmdiaz/HyperGraphs.jl): A Julia package for representing, analyzing, and generating hypergraphs which may be oriented and weighted.
+* [hyperG](https://cran.r-project.org/package=HyperG): An R package for storing and analyzing hypergraphs
+* [NetworkX](https://networkx.org): A Python package for representing, analyzing, and visualizing networks.
