### 2023-06-14 b1d891a
Fixed the _color_arg_to_dict function
