+++ tests/classes/test_simplicialcomplex.py
@@ -31,8 +31,8 @@ def test_constructor(edgelist5, dict5, incidence5, dataframe5):
     S_h = xgi.SimplicialComplex(H)
 
     assert set(S_h.nodes) == set(H.nodes) == set(S_list.nodes)
-    assert set(S_h.edges.members()) == set(S_list.edges.members())
-    assert set(H.edges.members()) <= set(S_h.edges.members()) # check it's a subset
+    assert S_h.edges.members() == S_list.edges.members()
+    assert H.edges.members() <= S_h.edges.members() # check it's a subset
 
 
 def test_string():
