### 2023-06-15 71d8498
fix: HG using SC as data (#399)
* fix: HG using SC as data

* test: add test for HG constructor with SC input data

* test: test for SC constructor with H input data + minor fix

* style: run black

* consistency in convert

*make convert_to_hypergraph consistent between constructor and standalone

*point to from_max_simplices in the docs

*adjust test

* addressing suggestions from max

* fix mistake in previous implementation
