+++ tests/classes/test_hypergraph.py
@@ -40,6 +40,12 @@ def test_constructor(edgelist5, dict5, incidence5, dataframe5):
     with pytest.raises(XGIError):
         xgi.Hypergraph(1)
 
+    SC = xgi.SimplicialComplex(edgelist5)
+    H_sc = xgi.Hypergraph(SC)
+
+    assert set(H_sc.nodes) == set(SC.nodes)
+    assert H_sc.edges.members() == SC.edges.members()
+
 
 def test_hypergraph_attrs():
     H = xgi.Hypergraph()
