### 2023-06-15 aee7db4
attempt at listing the available statistics
