### 2023-06-16 #115 @SimonMolinsky
Hi @nwlandry !

I've checked your package, and it is beyond expectations. Every point from the checklist is valid! I've included two comments that may be interesting for you (but both are optional, do as you wish :) ):

1. I didn't find the JOSS paper citation on your package webpage. Having the **Citation** header on your main page (additionally to the citation in the repo) would be nice. But it's up to you :) Usually, I cite packages by copying text from README or a documentation page.

2. This is 100% optional: you may consider using the **nbsphinx** plugin to render your tutorials into the documentation webpage. But it could be tedious because you must track changes in your tutorials and/or have these in two places, so I understand if you don't want to do it :)

**For now**:

I will start looking for the reviewers for your package. It should take around 1 week, hopefully not more :) @nwlandry  Could you help me a little with this task and provide **~5 categories** that your package falls into? Thanks in advance!
