### 2023-06-19 #115 @nwlandry
Categories that XGI falls into:
* network science
* higher-order networks (hypergraph, simplicial complex)
* data analysis and graph algorithms
* network visualization
* dataset I/O
