### 2023-06-19 29f961f
prefer to use S1.edges.members instead of relying on the fact that S1.edges.ids returns a dict
