### 2023-06-19 24b6c73
No longer necessary to check if self._ids is None since this is now dealth with in __init__. Also add tests and use check_like=True in an old test which ignores row order.
