### 2023-06-21 b5d41dd
Fix IDView.ids type (#406)
* IDView.ids should always return a set

* No longer necessary to check if self._ids is None since this is now dealt with in __init__. Also add tests and use check_like=True in an old test which ignores row order.
