### 2023-06-26 fc9f57f
slightly updated the acceptable args for color to dict
