### 2023-06-29 fae9512
changed single arg condition
