### 2023-06-30 1982067
small fix so that numbers of any type are accepted
