### 2023-06-30 087cb61
Fixed the _color_arg_to_dict function (#402)
* Fixed the _color_arg_to_dict function

* added unit tests

* Fixed docstring

* slightly updated the acceptable args for color to dict

* Response to review

* response to review part 2

* changed single arg condition

* small fix so that numbers of any type are accepted
