+++ xgi/drawing/draw.py
@@ -11,12 +11,13 @@ from matplotlib.colors import LinearSegmentedColormap, ListedColormap
 from matplotlib.patches import FancyArrow
 from mpl_toolkits.mplot3d.art3d import Line3DCollection, Poly3DCollection
 from networkx import spring_layout
+from numpy import ndarray
 from scipy.spatial import ConvexHull
 
 from .. import convert
 from ..classes import DiHypergraph, Hypergraph, SimplicialComplex, max_edge_order
 from ..exception import XGIError
-from ..stats import EdgeStat, NodeStat
+from ..stats import IDStat
 from .layout import _augmented_projection, barycenter_spring_layout
 
 __all__ = [
@@ -689,12 +690,12 @@ def draw_simplices(
     return ax
 
 
-def _scalar_arg_to_dict(arg, ids, min_val, max_val):
+def _scalar_arg_to_dict(scalar_arg, ids, min_val, max_val):
     """Map different types of arguments for drawing style to a dict with scalar values.
 
     Parameters
     ----------
-    arg : int, float, dict, iterable, or NodeStat/EdgeStat
+    scalar_arg : int, float, dict, iterable, or NodeStat/EdgeStat
         Attributes for drawing parameter.
     ids : NodeView or EdgeView
         This is the node or edge IDs that attributes get mapped to.
@@ -706,40 +707,89 @@ def _scalar_arg_to_dict(arg, ids, min_val, max_val):
     Returns
     -------
     dict
-        An ID: attribute dictionary.
+        An ID: scalar dictionary.
 
     Raises
     ------
     TypeError
         If a int, float, list, dict, or NodeStat/EdgeStat is not passed.
     """
-    if isinstance(arg, str):
+    if isinstance(scalar_arg, str):
         raise TypeError(
             "Argument must be int, float, dict, iterable, "
-            f"or NodeStat/EdgeStat. Received {type(arg)}"
+            f"or NodeStat/EdgeStat. Received {type(scalar_arg)}"
         )
-    elif isinstance(arg, dict):
-        return {id: arg[id] for id in arg if id in ids}
-    elif type(arg) in [int, float]:
-        return {id: arg for id in ids}
-    elif isinstance(arg, NodeStat) or isinstance(arg, EdgeStat):
-        vals = np.interp(arg.asnumpy(), [arg.min(), arg.max()], [min_val, max_val])
+
+    # Single argument
+    if isinstance(scalar_arg, (int, float)):
+        return {id: scalar_arg for id in ids}
+
+    # IDStat
+    if isinstance(scalar_arg, IDStat):
+        vals = np.interp(
+            scalar_arg.asnumpy(),
+            [scalar_arg.min(), scalar_arg.max()],
+            [min_val, max_val],
+        )
         return dict(zip(ids, vals))
-    elif isinstance(arg, Iterable):
-        return {id: arg[idx] for idx, id in enumerate(ids)}
-    else:
-        raise TypeError(
-            "Argument must be int, float, dict, iterable, "
-            f"or NodeStat/EdgeStat. Received {type(arg)}"
-        )
+
+    # Iterables of floats or ints
+    if isinstance(scalar_arg, Iterable):
+        if isinstance(scalar_arg, dict):
+            try:
+                return {id: float(scalar_arg[id]) for id in scalar_arg if id in ids}
+            except ValueError as e:
+                raise TypeError(
+                    "The input dict must have values that can be cast to floats."
+                )
+
+        elif isinstance(scalar_arg, (list, ndarray)):
+            try:
+                return {id: float(scalar_arg[idx]) for idx, id in enumerate(ids)}
+            except ValueError as e:
+                raise TypeError(
+                    "The input list or array must have values that can be cast to floats."
+                )
+        else:
+            raise TypeError(
[…]
