### 2023-07-07 686adb3
fix: improve structure of xgi.classes.function.complement
