### 2023-07-07 df528fb
fix: improve structure of xgi.classes.function.complement
