+++ tests/classes/test_function.py
@@ -793,6 +793,15 @@ def test_complement(edgelist1, edgelist2):
     H = xgi.Hypergraph(edgelist2)
     Hc = xgi.complement(H)
     assert Hc.edges.size.max() == 3
-    assert (
-        Hc.num_edges + H.num_edges == 41
-    )  # 41 is equal to choose(6, 1) + choose(6, 2) + choose(6, 3)
+    assert (Hc.num_edges + H.num_edges == 41)  # 41 = comb(6, 1) + comb(6, 2) + comb(6, 3)
+
+    H = xgi.empty_hypergraph()
+    H.add_edges_from([
+        ['a', 'b'],
+        ['b', 'c'],
+        ['c', 'd', 'e'],
+        ['e', 'a'],
+        ])
+    Hc = xgi.complement(H)
+    assert set(Hc.nodes) == set(H.nodes)
+    assert Hc.num_edges + H.num_edges == 25
