+++ xgi/utils/utilities.py
@@ -79,10 +79,14 @@ def dual_dict(edge_dict):
 
 
 def powerset(
-    iterable, include_empty=False, include_full=False, include_singletons=True, max_size=None
+    iterable,
+    include_empty=False,
+    include_full=False,
+    include_singletons=True,
+    max_size=None,
 ):
     """Returns all possible subsets of the elements in iterable, with options
-    to include the empty set and the set containing all elements, and to set 
+    to include the empty set and the set containing all elements, and to set
     the maximum subset size.
 
     Parameters
@@ -106,7 +110,7 @@ def powerset(
     -----
     include_empty overrides include_singletons if True: singletons will always
     be included if the empty set is.
-    Likewise, max_size will override other arguments: if set to -1, no subset 
+    Likewise, max_size will override other arguments: if set to -1, no subset
     will be returned.
 
     Examples
@@ -120,16 +124,14 @@ def powerset(
 
     start = 1 if include_singletons else 2
     start = 0 if include_empty else start  # overrides include_singletons if True
-    
+
     s = list(iterable)
-    
-    if max_size is None :
+
+    if max_size is None:
         max_size = len(s) if include_full else len(s) - 1
-    else :
+    else:
         max_size = min(max_size, len(s))
 
-
-    
     return chain.from_iterable(combinations(s, r) for r in range(start, max_size + 1))
 
 
