### 2023-07-08 54b3b56
rename reportviews to views and hypergraphviews to globalviews
