+++ docs/source/api/core/xgi.core.globalviews.rst
@@ -1,0 +1,10 @@
+﻿xgi.core.globalviews
+====================
+
+.. currentmodule:: xgi.core.globalviews
+
+.. automodule:: xgi.core.globalviews
+
+   .. rubric:: Functions
+
+   .. autofunction:: subhypergraph
