### 2023-07-08 74d523e
removed `is_empty` and `create_empty_copy`
