+++ xgi/algorithms/centrality.py
@@ -7,8 +7,8 @@ import numpy as np
 from numpy.linalg import norm
 from scipy.sparse.linalg import eigsh
 
-from ..core import convert_labels_to_integers, is_uniform
 from ..convert import to_line_graph
+from ..core import convert_labels_to_integers, is_uniform
 from ..exception import XGIError
 from ..linalg import clique_motif_matrix, incidence_matrix
 
