+++ tests/core/test_function.py
@@ -1,96 +1,7 @@
 import pytest
 
 import xgi
-from xgi.exception import IDNotFound, XGIError
-
-
-def test_num_edges_order(edgelist2):
-    H = xgi.Hypergraph(edgelist2)
-
-    assert xgi.num_edges_order(H, 0) == 0
-    assert xgi.num_edges_order(H, 1) == 2
-    assert xgi.num_edges_order(H, 2) == 1
-    assert xgi.num_edges_order(H) == 3
-
-
-def test_max_edge_order(edgelist1, edgelist4, edgelist5):
-    H0 = xgi.empty_hypergraph()
-    H1 = xgi.empty_hypergraph()
-    H1.add_nodes_from(range(5))
-    H2 = xgi.Hypergraph(edgelist1)
-    H3 = xgi.Hypergraph(edgelist4)
-    H4 = xgi.Hypergraph(edgelist5)
-
-    assert xgi.max_edge_order(H0) is None
-    assert xgi.max_edge_order(H1) == 0
-    assert xgi.max_edge_order(H2) == 2
-    assert xgi.max_edge_order(H3) == 3
-    assert xgi.max_edge_order(H4) == 3
-
-
-def test_is_possible_order(edgelist1):
-    H1 = xgi.Hypergraph(edgelist1)
-
-    assert xgi.is_possible_order(H1, -1) is False
-    assert xgi.is_possible_order(H1, 0) is True
-    assert xgi.is_possible_order(H1, 1) is True
-    assert xgi.is_possible_order(H1, 2) is True
-    assert xgi.is_possible_order(H1, 3) is False
-
-
-def test_is_uniform(edgelist1, edgelist6, edgelist7):
-    H0 = xgi.Hypergraph(edgelist1)
-    H1 = xgi.Hypergraph(edgelist6)
-    H2 = xgi.Hypergraph(edgelist7)
-    H3 = xgi.empty_hypergraph()
-
-    assert xgi.is_uniform(H0) is False
-    assert xgi.is_uniform(H1) == 2
-    assert xgi.is_uniform(H2) == 2
-    assert xgi.is_uniform(H3) is False
-
-
-def test_edge_neighborhood(edgelist3):
-    H = xgi.Hypergraph(edgelist3)
-    assert H.nodes.neighbors(3) == {1, 2, 4}
-    assert xgi.edge_neighborhood(H, 3) == [{1, 2}, {4}]
-    assert xgi.edge_neighborhood(H, 3, include_self=True) == [{1, 2, 3}, {3, 4}]
-    with pytest.raises(IDNotFound):
-        xgi.edge_neighborhood(H, 7)
-
-
-def test_degree_counts(edgelist1, edgelist2, edgelist3):
-    H1 = xgi.Hypergraph(edgelist1)
-    H2 = xgi.Hypergraph(edgelist2)
-    H3 = xgi.Hypergraph(edgelist3)
-
-    assert xgi.degree_counts(H1) == [0, 7, 1]
-    assert xgi.degree_counts(H2) == [0, 5, 1]
-    assert xgi.degree_counts(H3) == [0, 4, 2]
-
-    assert xgi.degree_counts(H1, order=2) == [2, 6]
-
-
-def test_degree_histogram(edgelist1, edgelist2, edgelist3):
-    H1 = xgi.Hypergraph(edgelist1)
-    H2 = xgi.Hypergraph(edgelist2)
-    H3 = xgi.Hypergraph(edgelist3)
-
-    assert xgi.degree_histogram(H1) == ([1, 2], [7, 1])
-    assert xgi.degree_histogram(H2) == ([1, 2], [5, 1])
[…]
