+++ xgi/algorithms/__init__.py
@@ -1,6 +1,7 @@
-from . import assortativity, centrality, clustering, connected
+from . import assortativity, centrality, clustering, connected, properties
 from .assortativity import *
 from .centrality import *
 from .clustering import *
 from .connected import *
 from .shortest_path import *
+from .properties import *
