+++ docs/source/api/core/xgi.core.function.rst
@@ -8,21 +8,10 @@
    .. rubric:: Functions
 
    .. autofunction:: convert_labels_to_integers
-   .. autofunction:: degree_counts
-   .. autofunction:: degree_histogram
-   .. autofunction:: density
-   .. autofunction:: edge_neighborhood
    .. autofunction:: freeze
    .. autofunction:: frozen
    .. autofunction:: get_edge_attributes
    .. autofunction:: get_node_attributes
-   .. autofunction:: incidence_density
    .. autofunction:: is_frozen
-   .. autofunction:: is_possible_order
-   .. autofunction:: is_uniform
-   .. autofunction:: max_edge_order
-   .. autofunction:: num_edges_order
    .. autofunction:: set_edge_attributes
    .. autofunction:: set_node_attributes
-   .. autofunction:: subfaces
-   .. autofunction:: unique_edge_sizes
