### 2023-07-08 aff8286
add freeze methods
