+++ xgi/core/function.py
@@ -10,212 +10,12 @@ from ..exception import IDNotFound, XGIError
 from .hypergraph import Hypergraph
 
 __all__ = [
-    "freeze",
-    "is_frozen",
     "set_node_attributes",
     "get_node_attributes",
     "set_edge_attributes",
     "get_edge_attributes",
     "convert_labels_to_integers",
 ]
-
-
-
-def degree_counts(H, order=None):
-    """Returns a list of the the number of occurrences of each degree value.
-
-    The counts correspond to degrees from 0 to max(degree).
-
-    Parameters
-    ----------
-    H : Hypergraph object
-        The hypergraph of interest
-    order: int, optional
-        Order of edges to take into account. If None (default),
-        consider all edges.
-
-    Returns
-    -------
-    list
-        A list of frequencies of degrees.
-        The degree values are the index in the list.
-
-    Notes
-    -----
-    Note: the bins are width one, hence len(list) can be large
-    (Order(num_edges))
-
-    The degree is defined as the number of edges to which a node belongs.
-    A node belonging only to a singleton edge will thus have degree 1 and
-    contribute accordingly to the degree count.
-
-    Examples
-    --------
-    >>> import xgi
-    >>> hyperedge_list = [[1, 2], [2, 3, 4]]
-    >>> H = xgi.Hypergraph(hyperedge_list)
-    >>> xgi.degree_counts(H)
-    [0, 3, 1]
-
-    """
-    counts = Counter(H.degree(order=order).values())
-    return [counts.get(i, 0) for i in range(max(counts) + 1)]
-
-
-def degree_histogram(H):
-    """Returns a degree histogram including bin centers (degree values).
-
-    Parameters
-    ----------
-    H : Hypergraph object
-        The hypergraph of interest
-
-    Returns
-    -------
-    tuple of lists
-        First entry is observed degrees (bin centers),
-            second entry is degree count (histogram height)
-    Notes
-    -----
-    Note: the bins are width one, hence there will be an entry
-    for every observed degree.
-
-    Examples
-    --------
-    >>> import xgi
-    >>> hyperedge_list = [[1, 2], [2, 3, 4]]
-    >>> H = xgi.Hypergraph(hyperedge_list)
-    >>> xgi.degree_histogram(H)
-    ([1, 2], [3, 1])
-
-    """
-    counts = Counter(H.degree().values())
-    degrees = []
-    heights = []
-    for d, c in sorted(counts.items(), key=lambda kv: kv[0]):
-        degrees.append(d)
-        heights.append(c)
-    return degrees, heights
-
-
-def unique_edge_sizes(H):
-    """A function that returns the unique edge sizes.
-
-    Parameters
-    ----------
-    H : Hypergraph object
-        The hypergraph of interest
-
-    Returns
-    -------
-    list()
-        The unique edge sizes in ascending order by size.
-    """
-    return sorted(set(H.edges.size.asnumpy()))
-
-
-def frozen(*args, **kwargs):
-    """Dummy method that raises an error when trying to modify frozen hypergraphs
-
-    Raises
-    ------
-    XGIError
-        Raises error when user tries to modify the hypergraph
-
-    Examples
-    --------
-    >>> import xgi
-    >>> hyperedge_list = [[1, 2], [2, 3, 4]]
-    >>> H = xgi.Hypergraph(hyperedge_list)
-    >>> xgi.freeze(H) # doctest: +ELLIPSIS
-    <xgi.core.hypergraph.Hypergraph object at 0x...>
-    >>> H.add_node(5)
-    Traceback (most recent call last):
-    xgi.exception.XGIError: Frozen hypergraph can't be modified
-
-    """
-    raise XGIError("Frozen hypergraph can't be modified")
-
-
-def freeze(H):
[…]
