### 2023-07-10 5678e93
Attempt at implementing suggestions by nich
