### 2023-07-10 d174113
second attempt at displaying the titles
