+++ docs/source/api/stats.rst
@@ -8,9 +8,8 @@ stats package
    
    The available statistics are:
 
-   ---------------
    Hypergraphs and simplicial complexes
-   ---------------
+   =====
 
    .. rubric:: Statistics of nodes
 
@@ -58,9 +57,8 @@ stats package
       ~xgi.stats.MultiDiNodeStat
       ~xgi.stats.MultiDiEdgeStat
 
-   ---------------
    Directed hypergraphs
-   ---------------
+   =====
    
 
    .. rubric:: Statistics of nodes in directed hypergraphs
