### 2023-07-10 87fffde
update of the panel for dihypergraph stats
