### 2023-07-13 xgi-org/xgi#414 @maximelucas
tutorials - new structure suggestion
In an effort to improved the tutorials, here is a proposed new structure:

- XGI in x minutes
  - in 1 min
  - in 5 min
  - in 15 min
- Workflows
  - Load dataset, clean, basic stats analysis and plots
  - Generate random structures, stats of the ensembles of realisations, plots
- Focus on one feature
  - drawing
  - stats
  - random generators
  - connected components
  - ...
- Case studies (reproduce results from papers - this could also be in Workflows?)
  - Zhang 2022
  - ...

If we like it, we could update this table of content with suggestions. And create separate issues to plan each tuto (if long enough to require that).
