### 2023-07-15 2b42e66
removed get methods, moved set methods to Hypergraph
