+++ xgi/core/hypergraph.py
@@ -355,8 +355,7 @@ class Hypergraph:
         See Also
         --------
         add_nodes_from
-        ~xgi.core.function.set_node_attributes
-        ~xgi.core.function.get_node_attributes
+        set_node_attributes
 
         Notes
         -----
@@ -386,8 +385,7 @@ class Hypergraph:
         See Also
         --------
         add_node
-        ~xgi.core.function.set_node_attributes
-        ~xgi.core.function.get_node_attributes
+        set_node_attributes
         """
         for n in nodes_for_adding:
             try:
@@ -463,6 +461,69 @@ class Hypergraph:
                 continue
             self.remove_node(n)
 
+    def set_node_attributes(self, values, name=None):
+        """Sets node attributes from a given value or dictionary of values.
+
+        Parameters
+        ----------
+        values : scalar value, dict-like
+            What the node attribute should be set to.  If `values` is
+            not a dictionary, then it is treated as a single attribute value
+            that is then applied to every node in `H`.  This means that if
+            you provide a mutable object, like a list, updates to that object
+            will be reflected in the node attribute for every node.
+            The attribute name will be `name`.
+
+            If `values` is a dict or a dict of dict, it should be keyed
+            by node to either an attribute value or a dict of attribute key/value
+            pairs used to update the node's attributes.
+        name : string, optional
+            Name of the node attribute to set if values is a scalar, by default None.
+
+        See Also
+        --------
+        set_edge_attributes
+        add_node
+        add_nodes_from
+
+        Notes
+        -----
+        After computing some property of the nodes of a hypergraph, you may
+        want to assign a node attribute to store the value of that property
+        for each node.
+
+        If you provide a list as the second argument, updates to the list
+        will be reflected in the node attribute for each node.
+
+        If you provide a dictionary of dictionaries as the second argument,
+        the outer dictionary is assumed to be keyed by node to an inner
+        dictionary of node attributes for that node.
+
+        Note that if the dictionary contains nodes that are not in `G`, the
+        values are silently ignored.
+
+        """
+        # Set node attributes based on type of `values`
+        if name is not None:  # `values` must not be a dict of dict
+            if isinstance(values, dict):  # `values` is a dict
+                for n, v in values.items():
+                    try:
+                        self._node_attr[n][name] = v
+                    except IDNotFound:
+                        warn(f"Node {n} does not exist!")
+            else:  # `values` is a constant
+                for n in self:
+                    self._node_attr[n][name] = values
+        else:  # `values` must be dict of dict
+            try:
+                for n, d in values.items():
+                    try:
+                        self._node_attr[n].update(d)
+                    except IDNotFound:
+                        warn(f"Node {n} does not exist!")
+            except (TypeError, ValueError, AttributeError):
+                raise XGIError("Must pass a dictionary of dictionaries")
+
     def add_edge(self, members, id=None, **attr):
         """Add one edge with optional attributes.
 
@@ -483,8 +544,7 @@ class Hypergraph:
         See Also
         --------
         add_edges_from : Add a collection of edges.
-        ~xgi.core.function.set_edge_attributes
-        ~xgi.core.function.get_edge_attributes
+        set_edge_attributes
 
         Examples
         --------
@@ -566,8 +626,7 @@ class Hypergraph:
         --------
         add_edge : Add a single edge.
         add_weighted_edges_from : Convenient way to add weighted edges.
-        ~xgi.core.function.set_edge_attributes
-        ~xgi.core.function.get_edge_attributes
+        set_edge_attributes
 
         Notes
         -----
@@ -763,6 +822,61 @@ class Hypergraph:
             )
         except KeyError:
             XGIError("Empty or invalid edges specified.")
+
[…]
