+++ xgi/algorithms/centrality.py
@@ -8,9 +8,9 @@ from numpy.linalg import norm
 from scipy.sparse.linalg import eigsh
 
 from ..convert import to_line_graph
-from ..utils import convert_labels_to_integers
 from ..exception import XGIError
 from ..linalg import clique_motif_matrix, incidence_matrix
+from ..utils import convert_labels_to_integers
 from .properties import is_uniform
 
 __all__ = [
