+++ docs/source/api/utils/xgi.utils.utilities.rst
@@ -24,3 +24,4 @@
    .. autofunction:: request_json_from_url
    .. autofunction:: request_json_from_url_cached
    .. autofunction:: subfaces
+   .. autofunction:: convert_labels_to_integers
