+++ docs/source/api/core/xgi.core.function.rst
@@ -1,14 +1,0 @@
-﻿xgi.core.function
-=================
-
-.. currentmodule:: xgi.core.function
-
-.. automodule:: xgi.core.function
-
-   .. rubric:: Functions
-
-   .. autofunction:: convert_labels_to_integers
-   .. autofunction:: get_edge_attributes
-   .. autofunction:: get_node_attributes
-   .. autofunction:: set_edge_attributes
-   .. autofunction:: set_node_attributes
