### 2023-07-16 xgi-org/xgi#414 @leotrs
I like these. What would we do with the existing tutorials?
