+++ docs/source/papers.rst
@@ -1,27 +1,0 @@
-XGI has been used in a variety of published work and other software.
-
-Published work
-==============
-
-2023
-----
-
-- `"Higher-order interactions shape collective dynamics differently in hypergraphs and simplicial complexes" <https://www.nature.com/articles/s41467-023-37190-9>`_
-
-Preprints
-=========
-
-2023
-----
-
-- `"Filtering higher-order datasets" <http://arxiv.org/abs/2305.06910>`_
-- `"Polarization in hypergraphs with community structure" <http://arxiv.org/abs/2302.13967>`_
-- Iac's new paper
-- Tim's new paper
-
-
-Other packages
---------------
-
-- simplicial-kuramoto
-- hypercontagion
