+++ docs/source/using-xgi.rst
@@ -1,0 +1,32 @@
+******************
+Projects using XGI
+******************
+
+XGI has been used in a variety of published work and other software.
+
+Published work
+==============
+
+2023
+----
+
+- Yuanzhao Zhang, Maxime Lucas, and Federico Battiston, `"Higher-order interactions shape collective dynamics differently in hypergraphs and simplicial complexes" <http://doi.org/10.1038/s41467-023-37190-9>`_, Nature Communications **14**, 1605 (2023).
+- Nicolo Ruggieri, Martina Contisciani, Federico Battiston, and Caterina de Bacco, `"Community detection in large hypergraphs" <https://doi.org/10.1126/sciadv.adg9159>`_, Science Advances **9**, eadg9159 (2023).
+
+Preprints
+=========
+
+2023
+----
+
+- Nicholas W. Landry, Ilya Amburg, Mirah Shi, and Sinan Aksoy, `"Filtering higher-order datasets" <http://arxiv.org/abs/2305.06910>`_, arXiv:2305.06910.
+- Nicholas W. Landry and Juan Restrepo, `"Polarization in hypergraphs with community structure" <http://arxiv.org/abs/2302.13967>`_, arXiv:2302.13967.
+- Iacopo Iacopini, Márton Karsai, and Alain Barrat, `"The temporal dynamics of group interactions in higher-order social networks" <https://arxiv.org/abs/2306.09967>`_, arXiv:2306.09967.
+- Timothy LaRock and Renaud Lambiotte, `"Encapsulation Structure and Dynamics in Hypergraphs" <https://arxiv.org/abs/2307.04613>`_, arXiv:2307.04613.
+
+
+Software packages
+=================
+
+- `simplicial-kuramoto <https://arnaudon.github.io/simplicial-kuramoto/>`_
+- `hypercontagion <https://hypercontagion.readthedocs.io/en/latest/>`_
