+++ README.md
@@ -11,6 +11,7 @@ Comple**X** **G**roup **I**nteractions (**XGI**) is a Python package for the rep
 * [**GitHub Discussions**](../../discussions)
 * [**Documentation**](https://xgi.readthedocs.io)
 * [**Contributors**](https://xgi.readthedocs.io/en/stable/contributors.html)
+* [**Projects using XGI**](https://xgi.readthedocs.io/en/stable/using-xgi.html)
 
 Sign up for our [mailing list](http://eepurl.com/igE6ez) and follow XGI on [Twitter](https://twitter.com/xginets) or [Mastodon](https://mathstodon.xyz/@xginets)!
 
