### 2023-07-17 #115 @SimonMolinsky
@JonnyTran , @Reckony

Please confirm your involvement here, and **copy & paste the review template** from here: [TEMPLATE](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template) into your responses. For now, it could be empty, you can go through it point by point and check empty boxes, or you can fill it when you end your review.

The review steps are described here: [Review Steps](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#a-guide-for-new-reviewers). If you feel that something needs to be clarified, I'll guide you!
