### 2023-07-17 546a16b
moving MultiDi-stats in the right place
