### 2023-07-17 0805882
fixes for issues raised
