+++ docs/source/index.rst
@@ -70,6 +70,7 @@ mind.
 - PyPI: `latest release <https://pypi.org/project/xgi/>`_
 - Twitter: `@xginets <https://twitter.com/xginets>`_
 - Contributors: `list <contributors.html>`_
+- Projects using XGI: `list <using-xgi.html`_
 
 Sign up for our `mailing list <http://eepurl.com/igE6ez>`_ and follow XGI on `Twitter <https://twitter.com/xginets>`_ or `Mastodon <https://mathstodon.xyz/@xginets>`_!
 
@@ -154,29 +155,6 @@ Academic References
   <https://www.nature.com/articles/s41567-019-0459-y>`_, Renaud Lambiotte, Martin
   Rosvall, and Ingo Scholtes.
 
-
-Contributors
-============
-
-The core XGI team members:
-
-* Nicholas Landry
-* Leo Torres
-* Maxime Lucas
-* Iacopo Iacopini
-* Giovanni Petri
-* Alice Patania
-* Alice Schwarze
-
-Other contributors:
-
-* Martina Contisciani
-* Tim LaRock
-* Sabina Adhikari
-* Marco Nurisso
-* Alexis Arnaudon
-* Thomas Robiglio
-* Gonzalo Contreras Aso
 
 Funding
 =======
