+++ xgi/convert.py
@@ -1,885 +1,0 @@
-from collections import defaultdict
-from copy import deepcopy
-from itertools import combinations
-
-import networkx as nx
-import pandas as pd
-from networkx.algorithms import bipartite
-from numpy import matrix, ndarray
-from scipy.sparse import (
-    coo_array,
-    coo_matrix,
-    csc_array,
-    csc_matrix,
-    csr_array,
-    csr_matrix,
-    lil_array,
-    lil_matrix,
-)
-
-from .classes import DiHypergraph, Hypergraph, SimplicialComplex, set_edge_attributes
-from .exception import XGIError
-from .generators import empty_dihypergraph, empty_hypergraph, empty_simplicial_complex
-from .linalg import adjacency_matrix, incidence_matrix
-
-__all__ = [
-    "convert_to_hypergraph",
-    "convert_to_dihypergraph",
-    "convert_to_graph",
-    "convert_to_simplicial_complex",
-    "from_hyperedge_list",
-    "to_hyperedge_list",
-    "from_hyperedge_dict",
-    "to_hyperedge_dict",
-    "from_bipartite_pandas_dataframe",
-    "to_bipartite_pandas_dataframe",
-    "from_incidence_matrix",
-    "from_max_simplices",
-    "to_incidence_matrix",
-    "from_bipartite_graph",
-    "to_bipartite_graph",
-    "dict_to_hypergraph",
-    "to_line_graph",
-    "from_simplex_dict",
-]
-
-
-def convert_to_hypergraph(data, create_using=None):
-    """Make a hypergraph from a known data structure.
-
-    The preferred way to call this is automatically from the class constructor.
-
-    Parameters
-    ----------
-    data : object to be converted
-        Current known types are:
-         * a Hypergraph object
-         * a SimplicialComplex object
-         * list-of-iterables
-         * dict-of-iterables
-         * Pandas DataFrame (bipartite edgelist)
-         * numpy matrix
-         * numpy ndarray
-         * scipy sparse matrix
-    create_using : Hypergraph constructor, optional (default=Hypergraph)
-        Hypergraph type to create. If hypergraph instance, then cleared before
-        populated.
-
-    Returns
-    -------
-    Hypergraph object
-        A hypergraph constructed from the data
-
-    See Also
-    --------
-    from_max_simplices : Constructs a hypergraph from the maximal simplices of a simplicial complex.
-
-    """
-    if data is None:
-        return empty_hypergraph(create_using)
-
-    elif isinstance(data, Hypergraph) and not isinstance(data, SimplicialComplex):
-        H = empty_hypergraph(create_using)
-        H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
-        ee = data.edges
-        H.add_edges_from((ee.members(e), e, deepcopy(attr)) for e, attr in ee.items())
-        H._hypergraph = deepcopy(data._hypergraph)
-        return H
-
-    elif isinstance(data, DiHypergraph):
-        H = empty_hypergraph(create_using)
-        H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
-        ee = data.edges
-        H.add_edges_from((ee.members(e), e, deepcopy(attr)) for e, attr in ee.items())
-        H._hypergraph = deepcopy(data._hypergraph)
-        if not isinstance(create_using, DiHypergraph):
-            return H
-
-    elif isinstance(data, SimplicialComplex):
-        H = empty_hypergraph(create_using)
-        H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
-        ee = data.edges
-        H.add_edges_from((ee.members(e), e, deepcopy(attr)) for e, attr in ee.items())
-        H._hypergraph = deepcopy(data._hypergraph)
[…]
