+++ docs/source/api/convert/xgi.convert.bipartite_graph.rst
@@ -1,0 +1,11 @@
+xgi.convert.bipartite_graph
+===========================
+
+.. currentmodule:: xgi.convert.bipartite_graph
+
+.. automodule:: xgi.convert.bipartite_graph
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: from_bipartite_graph
+   .. autofunction:: to_bipartite_graph
