+++ xgi/line_graphs.py
@@ -1,3 +1,5 @@
+from itertools import combinations
+
 import networkx as nx
 
 __all__ = [
