### 2023-07-24 d429758
Update docs/source/about.rst
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
