### 2023-07-24 85dd985
Update docs/source/using-xgi.rst
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
