### 2023-07-24 02e66f1
Added weighted option to to_line_graph and updated documentation.
