+++ xgi/line_graphs.py
@@ -7,14 +7,16 @@ __all__ = [
 ]
 
 
-def to_line_graph(H, s=1, normalize_weights=False):
+def to_line_graph(H, s=1, weighted=False, normalize_weights=False):
     """The s-line graph of the hypergraph.
 
-    The line graph of the hypergraph `H` is the graph whose
+    The s-line graph of the hypergraph `H` is the graph whose
     nodes correspond to each hyperedge in `H`, linked together
-    if they share at least one vertex. Edge weights correspond
-    to the size of the intersection between the hyperedges,
-    optionally normalized by the size of the smaller hyperedge.
+    if they share at least s vertices.
+
+    Optional edge weights correspond to the size of the
+    intersection between the hyperedges, optionally
+    normalized by the size of the smaller hyperedge.
 
     Parameters
     ----------
@@ -23,9 +25,13 @@ def to_line_graph(H, s=1, normalize_weights=False):
     s : int
         The intersection size to consider edges
         as connected, by default 1.
+    weighted : bool
+        If True, include edge weights corresponding
+        to the size of intersection between hyperedges.
     normalize_weights : bool
         If True, normalize edge weights by the
-        size of the smaller hyperedge.
+        size of the smaller hyperedge. Has no effect
+        if weighted=False.
 
     Returns
     -------
@@ -46,13 +52,20 @@ def to_line_graph(H, s=1, normalize_weights=False):
     LG.add_nodes_from(H.edges)
 
     for edge1, edge2 in combinations(H.edges.members(), 2):
+        # Check that the intersection size is larger than s
         intersection_size = len(edge1.intersection(edge2))
         if intersection_size >= s:
-            weight = intersection_size
-            if normalize_weights:
-                weight /= min([len(edge1), len(edge2)])
-
-            LG.add_edge(edge_label_dict[tuple(edge1)],
+            if not weighted:
+                # Add unweighted edge
+                LG.add_edge(edge_label_dict[tuple(edge1)],
+                            edge_label_dict[tuple(edge2)])
+            else:
+                # Compute the (normalized) weight
+                weight = intersection_size
+                if normalize_weights:
+                    weight /= min([len(edge1), len(edge2)])
+                # Add edge with weight
+                LG.add_edge(edge_label_dict[tuple(edge1)],
                         edge_label_dict[tuple(edge2)],
                         weight=weight)
 
