### 2023-07-24 97ca9c5
Update docs/source/using-xgi.rst
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
