### 2023-07-24 e623302
Update docs/source/using-xgi.rst
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
