### 2023-07-24 78b8ca4
Update docs/source/index.rst
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
