### 2023-07-24 cbe338b
Update docs/source/using-xgi.rst
Co-authored-by: Thomas Robiglio <83019028+thomasrobiglio@users.noreply.github.com>
