### 2023-07-24 6a461aa
Add papers and edit README (#422)
* rename

* initial commit

* added link in README

* Revert "rename"

This reverts commit 74ce6e379825654faab1aa64355617bf9a23428c.

* fixes for issues raised

* Modified the about page

* added buttons for github and papers

* Fix #424.

* Update docs/source/about.rst

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>

* Update docs/source/using-xgi.rst

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>

* Update docs/source/using-xgi.rst

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>

* Update docs/s
[…]
