+++ docs/source/using-xgi.rst
@@ -10,8 +10,15 @@ Published work
 2023
 ----
 
-- Yuanzhao Zhang, Maxime Lucas, and Federico Battiston, `"Higher-order interactions shape collective dynamics differently in hypergraphs and simplicial complexes" <http://doi.org/10.1038/s41467-023-37190-9>`_, Nature Communications **14**, 1605 (2023).
-- Nicolo Ruggieri, Martina Contisciani, Federico Battiston, and Caterina de Bacco, `"Community detection in large hypergraphs" <https://doi.org/10.1126/sciadv.adg9159>`_, Science Advances **9**, eadg9159 (2023).
+Yuanzhao Zhang, Maxime Lucas, and Federico Battiston, "Higher-order interactions shape collective dynamics differently in hypergraphs and simplicial complexes", *Nature Communications* **14**, 1605 (2023).
+
+:bdg-link-primary-line:`Paper <http://doi.org/10.1038/s41467-023-37190-9>`
+:bdg-link-primary-line:`Github <https://github.com/maximelucas/HOI_shape_sync_differently>`
+
+Nicolò Ruggieri, Martina Contisciani, Federico Battiston, and Caterina de Bacco, "Community detection in large hypergraphs", *Science Advances* **9**, eadg9159 (2023).
+
+:bdg-link-primary-line:`Paper <https://doi.org/10.1126/sciadv.adg9159>`
+:bdg-link-primary-line:`Github <https://github.com/nickruggeri/Hy-MMSBM>`
 
 Preprints
 =========
@@ -19,11 +26,24 @@ Preprints
 2023
 ----
 
-- Nicholas W. Landry, Ilya Amburg, Mirah Shi, and Sinan Aksoy, `"Filtering higher-order datasets" <http://arxiv.org/abs/2305.06910>`_, arXiv:2305.06910.
-- Nicholas W. Landry and Juan Restrepo, `"Polarization in hypergraphs with community structure" <http://arxiv.org/abs/2302.13967>`_, arXiv:2302.13967.
-- Iacopo Iacopini, Márton Karsai, and Alain Barrat, `"The temporal dynamics of group interactions in higher-order social networks" <https://arxiv.org/abs/2306.09967>`_, arXiv:2306.09967.
-- Timothy LaRock and Renaud Lambiotte, `"Encapsulation Structure and Dynamics in Hypergraphs" <https://arxiv.org/abs/2307.04613>`_, arXiv:2307.04613.
+Nicholas W. Landry, Ilya Amburg, Mirah Shi, and Sinan Aksoy, "Filtering higher-order datasets", arXiv:2305.06910 (2023).
 
+:bdg-link-primary-line:`Paper <http://arxiv.org/abs/2305.06910>`
+:bdg-link-primary-line:`Github <https://github.com/nwlandry/filtering-higher-order-datasets>`
+
+Nicholas W. Landry and Juan Restrepo, "Polarization in hypergraphs with community structure", arXiv:2302.13967 (2023).
+
+:bdg-link-primary-line:`Paper <http://arxiv.org/abs/2302.13967>`
+:bdg-link-primary-line:`Github <https://github.com/nwlandry/filtering-higher-order-datasets>`
+
+Iacopo Iacopini, Márton Karsai, and Alain Barrat, "The temporal dynamics of group interactions in higher-order social networks", arXiv:2306.09967 (2023).
+
+:bdg-link-primary-line:`Paper <https://arxiv.org/abs/2306.09967>`
+
+Timothy LaRock and Renaud Lambiotte, "Encapsulation Structure and Dynamics in Hypergraphs", arXiv:2307.04613 (2023).
+
[…]
