### 2023-07-24 ac73cd5
Refactor core data structures (#412)
* Rename `classes` to `core`

* rename reportviews to views and hypergraphviews to globalviews

* removed `is_empty` and `create_empty_copy`

* moved `subfaces` to `utils` module

* moved algorithms

* add freeze methods

* removed get methods, moved set methods to Hypergraph

* format with black

* move convert labels to int to utils

* update docs

* fix tests

* added unit tests
