+++ xgi/core/__init__.py
@@ -1,0 +1,4 @@
+from .hypergraph import Hypergraph
+from .dihypergraph import DiHypergraph
+from .globalviews import subhypergraph
+from .simplicialcomplex import SimplicialComplex
