+++ tests/classes/test_dihypergraph.py
@@ -1,491 +1,0 @@
-import pickle
-import tempfile
-
-import pytest
-
-import xgi
-from xgi.exception import IDNotFound, XGIError
-from xgi.utils import dual_dict
-
-
-def test_constructor(diedgelist1, diedgedict1):
-    H_list = xgi.DiHypergraph(diedgelist1)
-    H_dict = xgi.DiHypergraph(diedgedict1)
-    H_hg = xgi.DiHypergraph(H_list)
-
-    assert set(H_list.nodes) == set(H_dict.nodes) == set(H_hg.nodes)
-    assert set(H_list.edges) == set(H_dict.edges) == set(H_hg.edges)
-    for e in H_hg.edges:
-        assert (
-            H_list.edges.members(e) == H_dict.edges.members(e) == H_hg.edges.members(e)
-        )
-
-    with pytest.raises(XGIError):
-        xgi.DiHypergraph(1)
-
-
-def test_hypergraph_attrs():
-    H = xgi.DiHypergraph()
-    assert H._hypergraph == {}
-    with pytest.raises(XGIError):
-        H["name"]
-    H = xgi.DiHypergraph(name="test")
-    assert H["name"] == "test"
-
-
-def test_contains(diedgelist1):
-    el1 = diedgelist1
-    H = xgi.DiHypergraph(el1)
-    unique_nodes = {node for edge in el1 for node in edge[0].union(edge[1])}
-    for node in unique_nodes:
-        assert node in H
-
-    # test TypeError handling
-    assert [0] not in H
-
-
-def test_string():
-    H1 = xgi.DiHypergraph()
-    assert str(H1) == "Unnamed DiHypergraph with 0 nodes and 0 hyperedges"
-    H2 = xgi.DiHypergraph(name="test")
-    assert str(H2) == "DiHypergraph named test with 0 nodes and 0 hyperedges"
-
-
-def test_len(diedgelist1):
-    assert len(xgi.DiHypergraph(diedgelist1)) == 8
-
-
-def test_add_nodes_from(attr1, attr2, attr3):
-    H = xgi.DiHypergraph()
-    H.add_nodes_from(range(3), **attr1)
-    assert H.nodes[0]["color"] == attr1["color"]
-    assert H.nodes[1]["color"] == attr1["color"]
-    assert H.nodes[2]["color"] == attr1["color"]
-
-    H = xgi.DiHypergraph()
-    H.add_nodes_from(zip(range(3), [attr1, attr2, attr3]))
-    assert H.nodes[0]["color"] == attr1["color"]
-    assert H.nodes[1]["color"] == attr2["color"]
-    assert H.nodes[2]["color"] == attr3["color"]
-
-
-def test_add_node_attr(diedgelist1):
-    H = xgi.DiHypergraph(diedgelist1)
-    assert "new_node" not in H
-    H.add_node("new_node", color="red")
-    assert "new_node" in H
-    assert "color" in H.nodes["new_node"]
-    assert H.nodes["new_node"]["color"] == "red"
-
-
-def test_hypergraph_attr(diedgelist1):
-    H = xgi.DiHypergraph(diedgelist1)
-    with pytest.raises(XGIError):
-        H["color"]
-    H["color"] = "red"
-    assert H["color"] == "red"
-
-
-def test_memberships(diedgelist1):
-    H = xgi.DiHypergraph(diedgelist1)
-    assert H.nodes.memberships(1) == {0}
-    assert H.nodes.memberships(2) == {0}
-    assert H.nodes.memberships(3) == {0}
[…]
