+++ xgi/convert/higher_order_network.py
@@ -13,7 +13,7 @@ from scipy.sparse import (
     lil_matrix,
 )
 
-from ..classes import DiHypergraph, Hypergraph, SimplicialComplex
+from ..core import DiHypergraph, Hypergraph, SimplicialComplex
 from ..exception import XGIError
 from ..generators import empty_dihypergraph, empty_hypergraph, empty_simplicial_complex
 from .hyperedges import from_hyperedge_dict, from_hyperedge_list
