+++ docs/source/api/generators/xgi.generators.randomizing.rst
@@ -1,0 +1,11 @@
+xgi.generators.randomizing
+==========================
+
+.. currentmodule:: xgi.generators.randomizing
+
+.. automodule:: xgi.generators.randomizing
+   
+    .. rubric:: Functions
+   
+    .. autofunction:: shuffle_hyperedges
+    .. autofunction:: node_swap
