+++ xgi/generators/__init__.py
@@ -1,7 +1,16 @@
-from . import classic, uniform, random, simple, lattice, simplicial_complexes
+from . import (
+    classic,
+    uniform,
+    random,
+    simple,
+    lattice,
+    simplicial_complexes,
+    randomizing,
+)
 from .classic import *
 from .random import *
 from .uniform import *
 from .simple import *
 from .lattice import *
 from .simplicial_complexes import *
+from .randomizing import *
