### 2023-07-26 18eec96
feat: enforce equal aspect for circular layout #430
