### 2023-07-27 7fc3cee
changed implementation of set_aspect
