### 2023-07-27 949caec
feat: enforce equal aspect for circular layout #430 (#432)
* feat: enforce equal aspect for circular layout #430

* changed implementation of set_aspect
