+++ xgi/drawing/draw.py
@@ -1,5 +1,6 @@
 """Draw hypergraphs and simplicial complexes with matplotlib."""
 
+from collections import defaultdict
 from collections.abc import Iterable
 from inspect import signature
 from itertools import combinations
@@ -1159,12 +1160,16 @@ def draw_multilayer(
     H,
     pos=None,
     ax=None,
-    node_fc="tab:blue",
+    dyad_color="black",
+    dyad_lw=1.5,
+    edge_fc=None,
+    node_fc="white",
     node_ec="black",
-    node_lw=0.5,
-    node_size=5,
+    node_lw=1,
+    node_size=15,
     max_order=None,
-    palette="jet",
+    node_labels=False,
+    hyperedge_labels=False,
     conn_lines=True,
     conn_lines_style="dotted",
     width=5,
@@ -1172,6 +1177,7 @@ def draw_multilayer(
     h_angle=10,
     v_angle=0,
     sep=1,
+    **kwargs,
 ):
     """Draw a hypergraph or simplicial complex visualized in 3D
     showing hyperedges/simplices of different orders on superimposed layers.
@@ -1216,102 +1222,158 @@ def draw_multilayer(
     ax : matplotlib Axes3DSubplot
         The subplot with the multilayer network visualization.
     """
+    settings = {
+        "min_node_size": 10.0,
+        "max_node_size": 30.0,
+        "min_dyad_lw": 2.0,
+        "max_dyad_lw": 10.0,
+        "min_node_lw": 1.0,
+        "max_node_lw": 5.0,
+        "node_fc_cmap": cm.Reds,
+        "node_ec_cmap": cm.Greys,
+        "edge_fc_cmap": cm.Blues,
+        "dyad_color_cmap": cm.Greys,
+    }
+
+    settings.update(kwargs)
+
+    if edge_fc is None:
+        edge_fc = H.edges.size
 
     if pos is None:
         pos = barycenter_spring_layout(H)
 
     if ax is None:
-        fig, ax = plt.subplots(
+        _, ax = plt.subplots(
             1, 1, figsize=(width, height), dpi=600, subplot_kw={"projection": "3d"}
         )
 
     if max_order is None:
         max_order = max_edge_order(H)
-
-    cmap = cm.get_cmap(palette, max_order)
+    else:
+        max_order = min(max_order, max_edge_order(H))
 
     xs, ys = zip(*pos.values())
 
-    for order in range(1, max_order + 1):
-        zs = [order * sep] * len(xs)
+    dyad_color = _color_arg_to_dict(dyad_color, H.edges, settings["dyad_color_cmap"])
+    dyad_lw = _scalar_arg_to_dict(
+        dyad_lw, H.edges, settings["min_dyad_lw"], settings["max_dyad_lw"]
+    )
 
-        # draw lines connecting points on the different planes
-        if conn_lines and order > 1:
-            thru_nodes = H.nodes
-            lines3d_between = [
-                (list(pos[i]) + [order * sep - sep], list(pos[i]) + [order * sep])
-                for i in thru_nodes
-            ]
-            between_lines = Line3DCollection(
-                lines3d_between,
-                zorder=order,
-                color=".5",
-                alpha=0.4,
-                linestyle=conn_lines_style,
-                linewidth=1,
-            )
-            ax.add_collection3d(between_lines)
+    edge_fc = _color_arg_to_dict(edge_fc, H.edges, settings["edge_fc_cmap"])
 
-        # draw the edges/simplices of given order
-        edges = H.edges.filterby("order", order).members()
+    node_fc = _color_arg_to_dict(node_fc, H.nodes, settings["node_fc_cmap"])
[…]
