+++ xgi/drawing/draw.py
@@ -1,5 +1,6 @@
 """Draw hypergraphs and simplicial complexes with matplotlib."""
 
+from collections import defaultdict
 from collections.abc import Iterable
 from inspect import signature
 from itertools import combinations
@@ -112,8 +113,8 @@ def draw(
         If True, draw ids on the hyperedges. If a dict, must contain (edge_id: label)
         pairs.  By default, False.
     aspect : {"auto", "equal"} or float, optional
-        Set the aspect ratio of the axes scaling, i.e. y/x-scale. `aspect` is passed 
-        directly to matplotlib's `ax.set_aspect()`. Default is `equal`. See full 
+        Set the aspect ratio of the axes scaling, i.e. y/x-scale. `aspect` is passed
+        directly to matplotlib's `ax.set_aspect()`. Default is `equal`. See full
         description at
         https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.set_aspect.html
     **kwargs : optional args
@@ -1283,8 +1284,8 @@ def draw_hypergraph_hull(
     radius : float, optional
         Radius of the convex hull in the vicinity of the nodes, by default 0.05.
     aspect : {"auto", "equal"} or float, optional
-        Set the aspect ratio of the axes scaling, i.e. y/x-scale. `aspect` is passed 
-        directly to matplotlib's `ax.set_aspect()`. Default is `equal`. See full 
+        Set the aspect ratio of the axes scaling, i.e. y/x-scale. `aspect` is passed
+        directly to matplotlib's `ax.set_aspect()`. Default is `equal`. See full
         description at
         https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.set_aspect.html
     **kwargs : optional args
@@ -1419,12 +1420,16 @@ def draw_multilayer(
     H,
     pos=None,
     ax=None,
-    node_fc="tab:blue",
+    dyad_color="black",
+    dyad_lw=1.5,
+    edge_fc=None,
+    node_fc="white",
     node_ec="black",
-    node_lw=0.5,
-    node_size=5,
+    node_lw=1,
+    node_size=15,
     max_order=None,
-    palette="jet",
+    node_labels=False,
+    hyperedge_labels=False,
     conn_lines=True,
     conn_lines_style="dotted",
     width=5,
@@ -1432,6 +1437,7 @@ def draw_multilayer(
     h_angle=10,
     v_angle=0,
     sep=1,
+    **kwargs,
 ):
     """Draw a hypergraph or simplicial complex visualized in 3D
     showing hyperedges/simplices of different orders on superimposed layers.
@@ -1476,102 +1482,158 @@ def draw_multilayer(
     ax : matplotlib Axes3DSubplot
         The subplot with the multilayer network visualization.
     """
+    settings = {
+        "min_node_size": 10.0,
+        "max_node_size": 30.0,
+        "min_dyad_lw": 2.0,
+        "max_dyad_lw": 10.0,
+        "min_node_lw": 1.0,
+        "max_node_lw": 5.0,
+        "node_fc_cmap": cm.Reds,
+        "node_ec_cmap": cm.Greys,
+        "edge_fc_cmap": cm.Blues,
+        "dyad_color_cmap": cm.Greys,
+    }
+
+    settings.update(kwargs)
+
+    if edge_fc is None:
+        edge_fc = H.edges.size
 
     if pos is None:
         pos = barycenter_spring_layout(H)
 
     if ax is None:
-        fig, ax = plt.subplots(
+        _, ax = plt.subplots(
             1, 1, figsize=(width, height), dpi=600, subplot_kw={"projection": "3d"}
         )
 
     if max_order is None:
[…]
