+++ xgi/drawing/draw.py
@@ -20,6 +20,7 @@ from ..core import DiHypergraph, Hypergraph, SimplicialComplex
 from ..exception import XGIError
 from ..stats import IDStat
 from .layout import _augmented_projection, barycenter_spring_layout
+from .draw_utils import _scalar_arg_to_dict, _color_arg_to_dict, _CCW_sort, _draw_init, _update_lims
 
 __all__ = [
     "draw",
@@ -163,15 +164,7 @@ def draw(
     if edge_fc is None:
         edge_fc = H.edges.size
 
-    if pos is None:
-        pos = barycenter_spring_layout(H)
-
-    if ax is None:
-        ax = plt.gca()
-
-    ax.get_xaxis().set_ticks([])
-    ax.get_yaxis().set_ticks([])
-    ax.axis("off")
+    ax, pos = _draw_init(ax, pos)
 
     if not max_order:
         max_order = max_edge_order(H)
@@ -311,15 +304,7 @@ def draw_nodes(
 
     settings.update(kwargs)
 
-    if pos is None:
-        pos = barycenter_spring_layout(H)
-
-    if ax is None:
-        ax = plt.gca()
-
-    ax.get_xaxis().set_ticks([])
-    ax.get_yaxis().set_ticks([])
-    ax.axis("off")
+    ax, pos = _draw_init(ax, pos)
 
     # Note Iterable covers lists, tuples, ranges, generators, np.ndarrays, etc
     node_fc = _color_arg_to_dict(node_fc, H.nodes, settings["node_fc_cmap"])
@@ -436,21 +421,13 @@ def draw_hyperedges(
 
     """
 
-    if pos is None:
-        pos = barycenter_spring_layout(H)
-
-    if ax is None:
-        ax = plt.gca()
+    ax, pos = _draw_init(ax, pos)
 
     if max_order is None:
         max_order = max_edge_order(H)
 
     if edge_fc is None:
         edge_fc = H.edges.size
-
-    ax.get_xaxis().set_ticks([])
-    ax.get_yaxis().set_ticks([])
-    ax.axis("off")
 
     if settings is None:
         settings = {
@@ -604,18 +581,10 @@ def draw_simplices(
     if not max_order:
         max_order = max_edge_order(H_)
 
-    if pos is None:
-        pos = barycenter_spring_layout(H_)
-
-    if ax is None:
-        ax = plt.gca()
+    ax, pos = _draw_init(ax, pos)
 
     if edge_fc is None:
         edge_fc = H_.edges.size
-
-    ax.get_xaxis().set_ticks([])
-    ax.get_yaxis().set_ticks([])
-    ax.axis("off")
 
     if settings is None:
         settings = {
@@ -699,213 +668,7 @@ def draw_simplices(
     return ax
 
 
-def _scalar_arg_to_dict(scalar_arg, ids, min_val, max_val):
-    """Map different types of arguments for drawing style to a dict with scalar values.
 
-    Parameters
-    ----------
-    scalar_arg : int, float, dict, iterable, or NodeStat/EdgeStat
-        Attributes for drawing parameter.
-    ids : NodeView or EdgeView
-        This is the node or edge IDs that attributes get mapped to.
-    min_val : int or float
-        The minimum value of the drawing parameter.
-    max_val : int or float
-        The maximum value of the drawing parameter.
-
-    Returns
-    -------
-    dict
-        An ID: scalar dictionary.
-
-    Raises
-    ------
-    TypeError
-        If a int, float, list, dict, or NodeStat/EdgeStat is not passed.
-    """
-    if isinstance(scalar_arg, str):
-        raise TypeError(
-            "Argument must be int, float, dict, iterable, "
[…]
