### 2023-07-27 50d18d7
Refactor draw module (#435)
* refact: draw module

* style: black and isort
