+++ xgi/drawing/draw.py
@@ -19,6 +19,13 @@ from ..algorithms import max_edge_order
 from ..core import DiHypergraph, Hypergraph, SimplicialComplex
 from ..exception import XGIError
 from ..stats import IDStat
+from .draw_utils import (
+    _CCW_sort,
+    _color_arg_to_dict,
+    _draw_init,
+    _scalar_arg_to_dict,
+    _update_lims,
+)
 from .layout import _augmented_projection, barycenter_spring_layout
 
 __all__ = [
@@ -112,8 +119,8 @@ def draw(
         If True, draw ids on the hyperedges. If a dict, must contain (edge_id: label)
         pairs.  By default, False.
     aspect : {"auto", "equal"} or float, optional
-        Set the aspect ratio of the axes scaling, i.e. y/x-scale. `aspect` is passed 
-        directly to matplotlib's `ax.set_aspect()`. Default is `equal`. See full 
+        Set the aspect ratio of the axes scaling, i.e. y/x-scale. `aspect` is passed
+        directly to matplotlib's `ax.set_aspect()`. Default is `equal`. See full
         description at
         https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.set_aspect.html
     **kwargs : optional args
@@ -163,15 +170,7 @@ def draw(
     if edge_fc is None:
         edge_fc = H.edges.size
 
-    if pos is None:
-        pos = barycenter_spring_layout(H)
-
-    if ax is None:
-        ax = plt.gca()
-
-    ax.get_xaxis().set_ticks([])
-    ax.get_yaxis().set_ticks([])
-    ax.axis("off")
+    ax, pos = _draw_init(H, ax, pos)
 
     if not max_order:
         max_order = max_edge_order(H)
@@ -311,15 +310,7 @@ def draw_nodes(
 
     settings.update(kwargs)
 
-    if pos is None:
-        pos = barycenter_spring_layout(H)
-
-    if ax is None:
-        ax = plt.gca()
-
-    ax.get_xaxis().set_ticks([])
-    ax.get_yaxis().set_ticks([])
-    ax.axis("off")
+    ax, pos = _draw_init(H, ax, pos)
 
     # Note Iterable covers lists, tuples, ranges, generators, np.ndarrays, etc
     node_fc = _color_arg_to_dict(node_fc, H.nodes, settings["node_fc_cmap"])
@@ -436,21 +427,13 @@ def draw_hyperedges(
 
     """
 
-    if pos is None:
-        pos = barycenter_spring_layout(H)
-
-    if ax is None:
-        ax = plt.gca()
+    ax, pos = _draw_init(H, ax, pos)
 
     if max_order is None:
         max_order = max_edge_order(H)
 
     if edge_fc is None:
         edge_fc = H.edges.size
-
-    ax.get_xaxis().set_ticks([])
-    ax.get_yaxis().set_ticks([])
-    ax.axis("off")
 
     if settings is None:
         settings = {
@@ -604,18 +587,10 @@ def draw_simplices(
     if not max_order:
         max_order = max_edge_order(H_)
 
-    if pos is None:
-        pos = barycenter_spring_layout(H_)
-
-    if ax is None:
-        ax = plt.gca()
+    ax, pos = _draw_init(H_, ax, pos)
 
     if edge_fc is None:
         edge_fc = H_.edges.size
-
-    ax.get_xaxis().set_ticks([])
-    ax.get_yaxis().set_ticks([])
-    ax.axis("off")
 
     if settings is None:
         settings = {
@@ -697,215 +672,6 @@ def draw_simplices(
     _update_lims(pos, ax)
 
     return ax
-
-
-def _scalar_arg_to_dict(scalar_arg, ids, min_val, max_val):
[…]
