### 2023-07-27 0f7b2fa
Refactor draw module (#435)
* refact: draw module

* style: black and isort
