+++ xgi/drawing/draw_utils.py
@@ -1,0 +1,254 @@
+"""Draw hypergraphs and simplicial complexes with matplotlib."""
+
+from collections.abc import Iterable
+
+import matplotlib.pyplot as plt
+import numpy as np
+from matplotlib.colors import LinearSegmentedColormap, ListedColormap
+from numpy import ndarray
+
+from ..exception import XGIError
+from ..stats import IDStat
+from .layout import barycenter_spring_layout
+
+
+def _draw_init(H, ax, pos):
+
+    if pos is None:
+        pos = barycenter_spring_layout(H)
+
+    if ax is None:
+        ax = plt.gca()
+
+    ax.get_xaxis().set_ticks([])
+    ax.get_yaxis().set_ticks([])
+    ax.axis("off")
+
+    return ax, pos
+
+
+def _update_lims(pos, ax):
+    """Update Axis limits based on node positions"""
+
+    # compute axis limits
+    pos_arr = np.asarray([[x, y] for n, (x, y) in pos.items()])
+
+    maxx, maxy = np.max(pos_arr, axis=0)
+    minx, miny = np.min(pos_arr, axis=0)
+    w = maxx - minx
+    h = maxy - miny
+
+    # update view after drawing
+    padx, pady = 0.05 * w, 0.05 * h
+    corners = (minx - padx, miny - pady), (maxx + padx, maxy + pady)
+    ax.update_datalim(corners)
+    ax.autoscale_view()
+
+
+def _scalar_arg_to_dict(scalar_arg, ids, min_val, max_val):
+    """Map different types of arguments for drawing style to a dict with scalar values.
+
+    Parameters
+    ----------
+    scalar_arg : int, float, dict, iterable, or NodeStat/EdgeStat
+        Attributes for drawing parameter.
+    ids : NodeView or EdgeView
+        This is the node or edge IDs that attributes get mapped to.
+    min_val : int or float
+        The minimum value of the drawing parameter.
+    max_val : int or float
+        The maximum value of the drawing parameter.
+
+    Returns
+    -------
+    dict
+        An ID: scalar dictionary.
+
+    Raises
+    ------
+    TypeError
+        If a int, float, list, dict, or NodeStat/EdgeStat is not passed.
+    """
+    if isinstance(scalar_arg, str):
+        raise TypeError(
+            "Argument must be int, float, dict, iterable, "
+            f"or NodeStat/EdgeStat. Received {type(scalar_arg)}"
+        )
+
+    # Single argument
+    if isinstance(scalar_arg, (int, float)):
+        return {id: scalar_arg for id in ids}
+
+    # IDStat
+    if isinstance(scalar_arg, IDStat):
+        vals = np.interp(
+            scalar_arg.asnumpy(),
+            [scalar_arg.min(), scalar_arg.max()],
+            [min_val, max_val],
+        )
+        return dict(zip(ids, vals))
+
+    # Iterables of floats or ints
+    if isinstance(scalar_arg, Iterable):
+        if isinstance(scalar_arg, dict):
+            try:
+                return {id: float(scalar_arg[id]) for id in scalar_arg if id in ids}
+            except ValueError as e:
+                raise TypeError(
+                    "The input dict must have values that can be cast to floats."
+                )
+
+        elif isinstance(scalar_arg, (list, ndarray)):
+            try:
+                return {id: float(scalar_arg[idx]) for idx, id in enumerate(ids)}
+            except ValueError as e:
+                raise TypeError(
+                    "The input list or array must have values that can be cast to floats."
+                )
+        else:
+            raise TypeError(
+                "Argument must be an dict, list, or numpy array of floats or ints."
+            )
+
+    raise TypeError(
[…]
