+++ xgi/drawing/layout.py
@@ -22,7 +22,7 @@ __all__ = [
 ]
 
 
-def random_layout(H, center=None, dim=2, seed=None):
+def random_layout(H, center=None, seed=None):
     """Position nodes uniformly at random in the unit square.
 
     For every node, a position is generated by choosing each of dim coordinates
@@ -36,8 +36,6 @@ def random_layout(H, center=None, dim=2, seed=None):
     center : array-like, optional
         Coordinate pair around which to center the layout.
         If None (default), does not center the positions.
-    dim : int, optional
-        Dimension of layout, by default 2.
     seed : int, optional
         Set the random state for deterministic node layouts.
         If int, `seed` is the seed used by the random number generator,
@@ -76,15 +74,15 @@ def random_layout(H, center=None, dim=2, seed=None):
     if seed is not None:
         np.random.seed(seed)
 
-    H, center = nx.drawing.layout._process_params(H, center, dim)
-    pos = np.random.rand(len(H), dim) + center
+    H, center = nx.drawing.layout._process_params(H, center, 2)
+    pos = np.random.rand(len(H), 2) + center
     pos = pos.astype(np.float32)
     pos = dict(zip(H, pos))
 
     return pos
 
 
-def pairwise_spring_layout(H, seed=None):
+def pairwise_spring_layout(H, seed=None, k=None, **kwargs):
     """
     Position the nodes using Fruchterman-Reingold force-directed
     algorithm using the graph projection of the hypergraph
@@ -99,6 +97,13 @@ def pairwise_spring_layout(H, seed=None):
         If int, `seed` is the seed used by the random number generator,
         If None (default), random numbers are sampled from the
         numpy random number generator without initialization.
+    k : float
+        The spring constant of the links. When k=None (default),
+        k = 1/sqrt(N). For more information, see the documentation
+        for the NetworkX spring_layout() function.
+    kwargs :
+        Optional arguments for the NetworkX spring_layout() function.
+        See https://networkx.org/documentation/stable/reference/generated/networkx.drawing.layout.spring_layout.html
 
     Returns
     -------
@@ -131,7 +136,7 @@ def pairwise_spring_layout(H, seed=None):
     if isinstance(H, SimplicialComplex):
         H = convert.from_max_simplices(H)
     G = convert.convert_to_graph(H)
-    pos = nx.spring_layout(G, seed=seed)
+    pos = nx.spring_layout(G, seed=seed, k=k, **kwargs)
     return pos
 
 
@@ -190,7 +195,9 @@ def _augmented_projection(H, weighted=False):
     return G
 
 
-def barycenter_spring_layout(H, return_phantom_graph=False, seed=None):
+def barycenter_spring_layout(
+    H, return_phantom_graph=False, seed=None, k=None, **kwargs
+):
     """
     Position the nodes using Fruchterman-Reingold force-directed
     algorithm using an augmented version of the the graph projection
@@ -211,6 +218,14 @@ def barycenter_spring_layout(H, return_phantom_graph=False, seed=None):
         If int, `seed` is the seed used by the random number generator,
         If None (default), random numbers are sampled from the
         numpy random number generator without initialization.
+    k : float
+        The spring constant of the links. When k=None (default),
+        k = 1/sqrt(N). For more information, see the documentation
+        for the NetworkX spring_layout() function.
+    kwargs :
+        Optional arguments for the NetworkX spring_layout() function.
+        See https://networkx.org/documentation/stable/reference/generated/networkx.drawing.layout.spring_layout.html
+
 
     Returns
     -------
[…]
