### 2023-07-28 #115 @nwlandry
Hi @JonnyTran! Thanks for reaching out! From our end, there is no rush, so August 18th is fine. Best of luck with the interviews!
