### 2023-07-30 15003f7
streamline draw_multilayer function
