### 2023-07-30 1bef89d
updated defaults and docstring
