### 2023-07-31 51fbdc3
Added tests for weighted line graph implementation.
