### 2023-08-01 89e33b2
Added s=2 test.
