+++ tests/convert/test_line_graph.py
@@ -56,6 +56,15 @@ def test_abs_weighted_line_graph(edgelist1, hypergraph1, hypergraph2):
     assert L.edges[("e2","e3")]['weight'] == 1
     assert sum([dat["weight"] for _,_,dat in L.edges(data=True)]) == 3
 
+
+    L = xgi.to_line_graph(hypergraph1, s=2, weights="absolute")
+    assert isinstance(L, Graph)
+    assert all(["weight" in dat for u,v,dat in L.edges(data=True)])
+    assert set(L.nodes) == {"e1", "e2", "e3"}
+    assert [set(e) for e in L.edges] == [{"e1", "e2"}]
+    assert L.edges[("e1","e2")]['weight'] == 2
+    assert sum([dat["weight"] for _,_,dat in L.edges(data=True)]) == 2
+
     L = xgi.to_line_graph(hypergraph2, weights="absolute")
     assert isinstance(L, Graph)
     assert all(["weight" in dat for u,v,dat in L.edges(data=True)])
@@ -65,6 +74,15 @@ def test_abs_weighted_line_graph(edgelist1, hypergraph1, hypergraph2):
     assert L.edges[("e2","e3")]['weight'] == 2
     assert L.edges[("e1","e3")]['weight'] == 2
     assert sum([dat["weight"] for _,_,dat in L.edges(data=True)]) == 5
+
+    L = xgi.to_line_graph(hypergraph2, s=2, weights="absolute")
+    assert isinstance(L, Graph)
+    assert all(["weight" in dat for u,v,dat in L.edges(data=True)])
+    assert set(L.nodes) == {"e1", "e2", "e3"}
+    assert [set(e) for e in L.edges] == [{"e1", "e3"}, {"e2", "e3"}]
+    assert L.edges[("e2","e3")]['weight'] == 2
+    assert L.edges[("e1","e3")]['weight'] == 2
+    assert sum([dat["weight"] for _,_,dat in L.edges(data=True)]) == 4
 
 
 def test_normed_weighted_line_graph(edgelist1, hypergraph1, edgelist2, hypergraph2):
