### 2023-08-01 b17084e
Added encapsulation dag code.
