### 2023-08-01 2a33b6f
Added encapsulation dag tests.
