+++ xgi/convert/encapsulation_dag.py
@@ -1,5 +1,3 @@
-from itertools import combinations
-
 from ..exception import XGIError
 
 import networkx as nx
