### 2023-08-01 773f714
added cleanup method to all classes
