+++ xgi/core/dihypergraph.py
@@ -71,14 +71,14 @@ class DiHypergraph:
     Examples
     --------
     >>> import xgi
-    >>> H = xgi.DiHypergraph([([1, 2, 3], [4]), ([5, 6], [6, 7, 8])])
-    >>> H.nodes
+    >>> DH = xgi.DiHypergraph([([1, 2, 3], [4]), ([5, 6], [6, 7, 8])])
+    >>> DH.nodes
     DiNodeView((1, 2, 3, 4, 5, 6, 7, 8))
-    >>> H.edges
+    >>> DH.edges
     DiEdgeView((0, 1))
-    >>> [[sorted(h), sorted(t)] for h, t in H.edges.dimembers()]
+    >>> [[sorted(h), sorted(t)] for h, t in DH.edges.dimembers()]
     [[[1, 2, 3], [4]], [[5, 6], [6, 7, 8]]]
-    >>> [sorted(e) for e in H.edges.members()]
+    >>> [sorted(e) for e in DH.edges.members()]
     [[1, 2, 3, 4], [5, 6, 7, 8]]
     """
     _node_dict_factory = IDDict
@@ -247,8 +247,8 @@ class DiHypergraph:
             val = stat(*args, **kwargs).asdict()
             return val if node is None else val[node]
 
-        func.__doc__ = f"""Equivalent to H.{word}.{attr}.asdict(). For accepted *args and
-        **kwargs, see documentation of H.{word}.{attr}."""
+        func.__doc__ = f"""Equivalent to DH.{word}.{attr}.asdict(). For accepted *args and
+        **kwargs, see documentation of DH.{word}.{attr}."""
 
         return func
 
@@ -269,8 +269,8 @@ class DiHypergraph:
         --------
         >>> import xgi
         >>> hyperedge_list = [([1, 2], [2, 3, 4])]
-        >>> H = xgi.DiHypergraph(hyperedge_list)
-        >>> H.num_nodes
+        >>> DH = xgi.DiHypergraph(hyperedge_list)
+        >>> DH.num_nodes
         4
 
         """
@@ -293,8 +293,8 @@ class DiHypergraph:
         --------
         >>> import xgi
         >>> hyperedge_list = [([1, 2], [2, 3, 4])]
-        >>> H = xgi.DiHypergraph(hyperedge_list)
-        >>> H.num_edges
+        >>> DH = xgi.DiHypergraph(hyperedge_list)
+        >>> DH.num_edges
         1
         """
         return len(self._edge_in)
@@ -447,7 +447,7 @@ class DiHypergraph:
         values : scalar value, dict-like
             What the node attribute should be set to.  If `values` is
             not a dictionary, then it is treated as a single attribute value
-            that is then applied to every node in `H`.  This means that if
+            that is then applied to every node in `DH`.  This means that if
             you provide a mutable object, like a list, updates to that object
             will be reflected in the node attribute for every node.
             The attribute name will be `name`.
@@ -531,9 +531,9 @@ class DiHypergraph:
         Add edges with or without specifying an edge id.
 
         >>> import xgi
-        >>> H = xgi.DiHypergraph()
-        >>> H.add_edge(([1, 2, 3], [2, 3, 4]))
-        >>> H.add_edge(([3, 4], set()), id='myedge')
+        >>> DH = xgi.DiHypergraph()
+        >>> DH.add_edge(([1, 2, 3], [2, 3, 4]))
+        >>> DH.add_edge(([3, 4], set()), id='myedge')
         """
         if not members:
             raise XGIError("Cannot add an empty edge")
@@ -624,57 +624,57 @@ class DiHypergraph:
         Examples
         --------
         >>> import xgi
-        >>> H = xgi.DiHypergraph()
+        >>> DH = xgi.DiHypergraph()
 
[…]
