### 2023-08-01 4503e9c
feat: return mappable to allow colorbar
