### 2023-08-01 038573d
added more to the cleanup method
