### 2023-08-01 8e3eaf1
refact: simplified and improved draw_nodes by fully using plt.scatter
