### 2023-08-02 0b19dc1
Added weights option to to_line_graph function. (#427)
* Added weights option to to_line_graph function.

* Added tests for weighted line graph implementation.

* Added s=2 test.
