### 2023-08-02 7aca0e0
docs: updated docstrings for draw_nodes. Removed node_ec_cmap
