### 2023-08-02 a8070dc
fix: remove unused code
