### 2023-08-03 9d90532
Refactored relations to subset_types. Exposed and documented empirical relations filtering function. Documentation improvements.
