### 2023-08-03 93768d7
`remove_node` and `remove_nodes_from` now inherit from `Hypergraph`
