### 2023-08-03 5e6f475
added tests for SimplicialComplex and DiHypergraph
