### 2023-08-03 b0b01e7
`is_frozen` now a property
