### 2023-08-03 52e6aac
Refactor multilayer drawing function and add headers to convert module (#434)
* initial commit

* Refactor draw module (#435)

* refact: draw module

* style: black and isort

* fix #331 (#438)

* Added the ability for users to access the optional arguments of NetworkX layout functions. (#439)

* initial commit

* remove order_nodes

* streamline draw_multilayer function

* updated defaults and docstring

* added brief convert descriptions

* response to review

---------

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
