+++ xgi/drawing/draw.py
@@ -1,24 +1,20 @@
 """Draw hypergraphs and simplicial complexes with matplotlib."""
 
-from collections.abc import Iterable
 from inspect import signature
 from itertools import combinations
 
 import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib import cm
-from matplotlib.colors import LinearSegmentedColormap, ListedColormap
 from matplotlib.patches import FancyArrow
 from mpl_toolkits.mplot3d.art3d import Line3DCollection, Poly3DCollection
 from networkx import spring_layout
-from numpy import ndarray
 from scipy.spatial import ConvexHull
 
 from .. import convert
-from ..algorithms import max_edge_order
+from ..algorithms import max_edge_order, unique_edge_sizes
 from ..core import DiHypergraph, Hypergraph, SimplicialComplex
 from ..exception import XGIError
-from ..stats import IDStat
 from .draw_utils import (
     _CCW_sort,
     _color_arg_to_dict,
@@ -119,6 +115,8 @@ def draw(
         If True, draw ids on the hyperedges. If a dict, must contain (edge_id: label)
         pairs.  By default, False.
     aspect : {"auto", "equal"} or float, optional
+        Set the aspect ratio of the axes scaling, i.e. y/x-scale. `aspect` is passed
+        directly to matplotlib's `ax.set_aspect()`. Default is `equal`. See full
         Set the aspect ratio of the axes scaling, i.e. y/x-scale. `aspect` is passed
         directly to matplotlib's `ax.set_aspect()`. Default is `equal`. See full
         description at
@@ -1159,19 +1157,23 @@ def draw_multilayer(
     H,
     pos=None,
     ax=None,
-    node_fc="tab:blue",
+    dyad_color="black",
+    dyad_lw=0.5,
+    edge_fc=None,
+    node_fc="white",
     node_ec="black",
     node_lw=0.5,
     node_size=5,
+    plane_color="grey",
     max_order=None,
-    palette="jet",
     conn_lines=True,
     conn_lines_style="dotted",
     width=5,
     height=5,
     h_angle=10,
-    v_angle=0,
+    v_angle=20,
     sep=1,
+    **kwargs,
 ):
     """Draw a hypergraph or simplicial complex visualized in 3D
     showing hyperedges/simplices of different orders on superimposed layers.
@@ -1181,21 +1183,57 @@ def draw_multilayer(
     H : Hypergraph or SimplicialComplex.
         Higher-order network to plot.
     pos : dict or None, optional
-        The positions of the nodes in the multilayer network. If None, a default layout will be computed using xgi.barycenter_spring_layout(). Default is None.
+        The positions of the nodes in the multilayer network.
+        If None, a default layout will be computed using
+        xgi.barycenter_spring_layout(). Default is None.
     ax : matplotlib Axes3DSubplot or None, optional
-        The subplot to draw the visualization on. If None, a new subplot will be created. Default is None.
-    node_fc : color or sequence of colors, optional
-        The face color(s) of the nodes. Default is "tab:blue".
-    node_ec : color or sequence of colors, optional
-        The edge color(s) of the nodes. Default is "black".
-    node_lw : float or sequence of floats, optional
-        The linewidth(s) of the node edges. Default is 0.5.
-    node_size : scalar or array-like, optional
-        The size(s) of the nodes. Default is 5.
-    max_order : int or None, optional
-        The maximum order of hyperedges/simplices to consider for coloring. If None edges up to the maximal order are drawn. Default is None.
-    palette : str, optional
-        The name of the matplotlib color palette to use. Default is 'jet'.
+        The subplot to draw the visualization on.
+        If None, a new subplot will be created. Default is None.
[…]
