### 2023-08-04 c222384
Update xgi/convert/encapsulation_dag.py documentation
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
