### 2023-08-04 b68ef6e
Fix typo in xgi/convert/encapsulation_dag.py
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
