+++ xgi/convert/encapsulation_dag.py
@@ -5,7 +5,8 @@ import networkx as nx
 __all__ = ["to_encapsulation_dag", "empirical_subsets_filter"]
 
 def to_encapsulation_dag(H, subset_types="all"):
-    """The encapsulation DAG of the hypergraph.
+    """The encapsulation DAG (Directed Acyclic Graph) of
+    the hypergraph H.
 
     An encapsulation DAG is a directed line graph
     where the nodes are hyperedges in H and a directed edge
@@ -21,15 +22,33 @@ def to_encapsulation_dag(H, subset_types="all"):
             "all" : all subset relationships
             "immediate" : only subset relationships between hyperedges of
                 adjacent sizes (i.e., edges from k to k-1)
-            "empirical" : only subset relationships between hyperedges of
-                minimum size difference k and k'. A relaxation of the
-                "immediate" option. For example, a hyperedge of size 4 may
-                have no immediate encapsulation relationships with hyperedges
-                of size 3, but may encapsulate hyperedegs of size 2.
+            "empirical" : A relaxation of the "immediate" option where only
+                subset relationships between hyperedges of size k and subsets
+                of maximum size k'<k existing in the hypergraph are included.
+                For example, a hyperedge of size 5 may have no immediate
+                encapsulation relationships with hyperedges of size 4, but may
+                encapsulate hyperedegs of size 3, which will be included if
+                using this setting (whereas relationships with subsets of size 2
+                would not be included).
     Returns
     -------
     LG : networkx.DiGraph
          The line graph associated to the Hypergraph
+
+    Examples
+    --------
+    >>> import xgi
+    >>> from xgi.convert import to_encapsulation_dag, empirical_subsets_filter
+    >>> H = xgi.Hypergraph([["a","b","c"], ["b","c","f"], ["a","b"], ["c", "e"], ["a"], ["f"]])
+    >>> dag = to_encapsulation_dag(H)
+    >>> dag.edges()
+    OutEdgeView([(0, 2), (0, 4), (2, 4), (1, 5)])
+    >>> dag = to_encapsulation_dag(H, subset_types="immediate")
+    >>> dag.edges()
+    OutEdgeView([(0, 2), (2, 4)])
+    >>> dag = to_encapsulation_dag(H, subset_types="empirical")
+    >>> dag.edges()
+    OutEdgeView([(0, 2), (2, 4), (1, 5)])
 
     References
     ----------
@@ -37,7 +56,6 @@ def to_encapsulation_dag(H, subset_types="all"):
     & Renaud Lambiotte. https://arxiv.org/abs/2307.04613
 
     """
-
     if not (subset_types in ["all", "immediate", "empirical"]):
         raise XGIError(f"{subset_types} not a valid subset_types option. Choices are "
                        "'all', 'immediate', and 'empirical'.")
@@ -72,29 +90,77 @@ def to_encapsulation_dag(H, subset_types="all"):
 
     return dag
 
-
 def _encapsulated(larger, smaller):
+    """Test if a larger hyperedge encapsulates a smaller hyperedge.
+
+    Parameters
+    ----------
+    larger : Set
+        larger hyperedge
+    smaller : Set
+        smaller hyperedge
+    Returns
+    -------
+    _ : Bool
+        True if the size of the intersection between larger and smaller is
+            the same as the size of smaller.
+
+    """
     return len(set(larger).intersection(set(smaller))) == len(smaller)
 
-def _get_candidates(relations, H, he):
+def _get_candidates(subset_types, H, he):
+    """Get the candidate hyperedges for encapsulation by he based on the subset type.
+
+    Parameters
+    ----------
+    subset_types : str
+        Type of subset relationships
+    H : Hypergraph
+        The hypergraph
+    he : Set
+        The hyperedge
+    Returns
+    -------
+    candidates : Set
[…]
