### 2023-08-04 ecc464f
Minor refactor.
