+++ xgi/convert/encapsulation_dag.py
@@ -1,0 +1,208 @@
+from ..exception import XGIError
+
+import networkx as nx
+
+__all__ = ["to_encapsulation_dag", "empirical_subsets_filter"]
+
+def to_encapsulation_dag(H, subset_types="all"):
+    """The encapsulation DAG (Directed Acyclic Graph) of
+    the hypergraph H.
+
+    An encapsulation DAG is a directed line graph
+    where the nodes are hyperedges in H and a directed edge
+    exists from a larger hyperedge to a smaller hyperedge if
+    the smaller is a subset of the larger.
+
+    Parameters
+    ----------
+    H : Hypergraph
+        The hypergraph of interest
+    subset_types : str, optional
+        Type of relations to include. Options are:
+            "all" : all subset relationships
+            "immediate" : only subset relationships between hyperedges of
+                adjacent sizes (i.e., edges from k to k-1)
+            "empirical" : A relaxation of the "immediate" option where only
+                subset relationships between hyperedges of size k and subsets
+                of maximum size k'<k existing in the hypergraph are included.
+                For example, a hyperedge of size 5 may have no immediate
+                encapsulation relationships with hyperedges of size 4, but may
+                encapsulate hyperedegs of size 3, which will be included if
+                using this setting (whereas relationships with subsets of size 2
+                would not be included).
+    Returns
+    -------
+    LG : networkx.DiGraph
+         The line graph associated to the Hypergraph
+
+    Examples
+    --------
+    >>> import xgi
+    >>> from xgi.convert import to_encapsulation_dag, empirical_subsets_filter
+    >>> H = xgi.Hypergraph([["a","b","c"], ["b","c","f"], ["a","b"], ["c", "e"], ["a"], ["f"]])
+    >>> dag = to_encapsulation_dag(H)
+    >>> dag.edges()
+    OutEdgeView([(0, 2), (0, 4), (2, 4), (1, 5)])
+    >>> dag = to_encapsulation_dag(H, subset_types="immediate")
+    >>> dag.edges()
+    OutEdgeView([(0, 2), (2, 4)])
+    >>> dag = to_encapsulation_dag(H, subset_types="empirical")
+    >>> dag.edges()
+    OutEdgeView([(0, 2), (2, 4), (1, 5)])
+
+    References
+    ----------
+    "Encapsulation Structure and Dynamics in Hypergraphs", by Timothy LaRock
+    & Renaud Lambiotte. https://arxiv.org/abs/2307.04613
+
+    """
+    if not (subset_types in ["all", "immediate", "empirical"]):
+        raise XGIError(f"{subset_types} not a valid subset_types option. Choices are "
+                       "'all', 'immediate', and 'empirical'.")
+
+    # Construct the dag
+    dag = nx.DiGraph()
+    # Loop over hyperedges
+    for he_idx in H.edges:
+        # Add the hyperedge as a node
+        dag.add_node(he_idx)
+        # Get the hyperedge as a set
+        he = H.edges.members(he_idx)
+        # Get candidate encapsulation hyperedges
+        candidates = _get_candidates(subset_types, H, he)
+        # for each candidate
+        candidates_checked = set()
+        for cand_idx in candidates:
+            if cand_idx in candidates_checked:
+                continue
+            cand = H.edges.members(cand_idx)
+            if len(he) > len(cand):
+                if _encapsulated(he, cand):
+                    dag.add_edge(he_idx, cand_idx)
+            elif len(cand) > len(he):
+                if _encapsulated(cand, he):
+                    dag.add_edge(cand_idx, he_idx)
+
+    # If empirically closest subsets, filter out all edges except those
+    # between k and maximum existing k'<k
+    if subset_types == "empirical":
+        empirical_subsets_filter(H, dag)
+
+    return dag
+
[…]
