+++ tests/core/test_hypergraph.py
@@ -536,6 +536,10 @@ def test_issue_445(edgelist1):
     assert 0 not in H.edges
     assert H._edge == xgi.dual_dict(H._node)
 
+    H = xgi.Hypergraph([[1, 2, 3], [1, 2]])
+    H.remove_node(1, strong=True)
+    assert H._edge == xgi.dual_dict(H._node)
+
 
 def test_clear_edges(edgelist1):
     H = xgi.Hypergraph(edgelist1)
