### 2023-08-04 ab2a2c7
Added cleanup method to all classes (#440)
* added cleanup method to all classes

* added more to the cleanup method

* `remove_node` and `remove_nodes_from` now inherit from `Hypergraph`

* fix 445 and add test

* added tests for SimplicialComplex and DiHypergraph

* lots of tests!!

* small fix

* small fix

* add tests

* `is_frozen` now a property

* updated docs

* added test
