### 2023-08-11 1d26dfa
Changed rho to an optional arg in HPPM (#446)
