### 2023-08-11 290729a
Fix empty edge error when loading BiGG data. (#447)
* Fix error when loading BiGG data

* format: isort and black
