### 2023-08-16 2e1e715
feat: accept dict as input in draw
