### 2023-08-16 f53a463
add check for negative input values #272
