### 2023-08-16 e832f36
feat: added node_shape argument in draw_nodes
