### 2023-08-16 88661c3
docs: added colorbar note
