### 2023-08-16 e241d00
tests: fixed and added more. fix: now plotting nodes with non-finite color. warning: removed when cmap is ignoed because single color
