+++ tests/drawing/test_draw.py
@@ -172,7 +172,16 @@ def test_draw_nodes(edgelist8):
     fig, ax = plt.subplots()
     ax, node_collection = xgi.draw_nodes(H, ax=ax)
     fig2, ax2 = plt.subplots()
-    ax2, node_collection2 = xgi.draw_nodes(H, ax=ax2, node_fc="r", node_ec="b", node_lw=2, node_size=20, zorder=10, node_shape="v")
+    ax2, node_collection2 = xgi.draw_nodes(
+        H,
+        ax=ax2,
+        node_fc="r",
+        node_ec="b",
+        node_lw=2,
+        node_size=20,
+        zorder=10,
+        node_shape="v",
+    )
 
     # number of elements
     assert len(ax.lines) == 0  # dyads
@@ -181,25 +190,32 @@ def test_draw_nodes(edgelist8):
     assert offsets.shape[0] == H.num_nodes  # nodes
 
     # zorder
-    assert node_collection.get_zorder() == 0 
-    assert node_collection2.get_zorder() == 10  
+    assert node_collection.get_zorder() == 0
+    assert node_collection2.get_zorder() == 10
 
     # facecolor
-    assert np.all(node_collection.get_facecolor() == np.array([[1., 1., 1., 1.]])) # white
-    assert np.all(node_collection2.get_facecolor() == np.array([[1., 0., 0., 1.]])) # blue
+    assert np.all(
+        node_collection.get_facecolor() == np.array([[1.0, 1.0, 1.0, 1.0]])
+    )  # white
+    assert np.all(
+        node_collection2.get_facecolor() == np.array([[1.0, 0.0, 0.0, 1.0]])
+    )  # blue
 
     # facecolor
-    assert np.all(node_collection.get_edgecolor() == np.array([[0., 0., 0., 1.]])) # black
-    assert np.all(node_collection2.get_edgecolor() == np.array([[0., 0., 1., 1.]])) # red
+    assert np.all(
+        node_collection.get_edgecolor() == np.array([[0.0, 0.0, 0.0, 1.0]])
+    )  # black
+    assert np.all(
+        node_collection2.get_edgecolor() == np.array([[0.0, 0.0, 1.0, 1.0]])
+    )  # red
 
     # linewidth
-    assert np.all(node_collection.get_linewidth() == np.array([1])) 
+    assert np.all(node_collection.get_linewidth() == np.array([1]))
     assert np.all(node_collection2.get_linewidth() == np.array([2]))
 
     # linewidth
-    assert np.all(node_collection.get_sizes() == np.array([15**2])) 
+    assert np.all(node_collection.get_sizes() == np.array([15**2]))
     assert np.all(node_collection2.get_sizes() == np.array([20**2]))
-
 
     plt.close("all")
 
