### 2023-08-16 a9e81ea
fix: raise for negative node size
