+++ xgi/drawing/draw.py
@@ -356,6 +356,10 @@ def draw_nodes(
         node_size = node_size.asnumpy()
     elif isinstance(node_size, dict):
         node_size = np.array(list(node_size.values()))
+    # check validity of input values
+    if np.any(node_size < 0):
+        raise ValueError("node_size cannot contain negative values.")
+    #clip 
     node_size = np.clip(node_size, settings["min_node_size"], settings["max_node_size"])
     node_size = node_size**2
 
@@ -368,13 +372,12 @@ def draw_nodes(
         node_lw = node_lw.asnumpy()
     elif isinstance(node_lw, dict):
         node_lw = np.array(list(node_lw.values()))
-    node_lw = np.clip(node_lw, settings["min_node_lw"], settings["max_node_lw"])
-
     # check validity of input values
-    if np.any(node_size < 0):
-        raise ValueError("node_size cannot contain negative values.")
     if np.any(node_lw < 0):
         raise ValueError("node_lw cannot contain negative values.")
+    # clip 
+    node_lw = np.clip(node_lw, settings["min_node_lw"], settings["max_node_lw"])
+
 
     # plot
     node_collection = ax.scatter(
