### 2023-08-16 fd47b34
fix: raise for negative node lw
