### 2023-08-16 3c72b5e
test: added for cmap and vmin/vmax in draw_nodes
