### 2023-08-17 ba2d28e
Added `s` parameter to `neighbors`
