### 2023-08-17 f513dfa
Add a recipe to the cookbook
