### 2023-08-18 ed4e728
docs: added docstring to update_lims
