### 2023-08-18 7c5826e
refact: new _draw_arg_to_arr function
