+++ tests/drawing/test_draw.py
@@ -166,101 +166,7 @@ def test_draw(edgelist8):
     plt.close()
 
 
-def test_draw_nodes(edgelist8):
-    H = xgi.Hypergraph(edgelist8)
 
-    fig, ax = plt.subplots()
-    ax, node_collection = xgi.draw_nodes(H, ax=ax)
-    fig2, ax2 = plt.subplots()
-    ax2, node_collection2 = xgi.draw_nodes(
-        H,
-        ax=ax2,
-        node_fc="r",
-        node_ec="b",
-        node_lw=2,
-        node_size=20,
-        zorder=10,
-        node_shape="v",
-    )
-
-    # number of elements
-    assert len(ax.lines) == 0  # dyads
-    assert len(ax.patches) == 0  # hyperedges
-    offsets = node_collection.get_offsets()
-    assert offsets.shape[0] == H.num_nodes  # nodes
-
-    # node_fc
-    assert np.all(
-        node_collection.get_facecolor() == np.array([[1.0, 1.0, 1.0, 1.0]])
-    )  # white
-    assert np.all(
-        node_collection2.get_facecolor() == np.array([[1.0, 0.0, 0.0, 1.0]])
-    )  # blue
-
-    # node_ec
-    assert np.all(
-        node_collection.get_edgecolor() == np.array([[0.0, 0.0, 0.0, 1.0]])
-    )  # black
-    assert np.all(
-        node_collection2.get_edgecolor() == np.array([[0.0, 0.0, 1.0, 1.0]])
-    )  # red
-
-    # node_lw
-    assert np.all(node_collection.get_linewidth() == np.array([1]))
-    assert np.all(node_collection2.get_linewidth() == np.array([2]))
-
-    # node_size
-    assert np.all(node_collection.get_sizes() == np.array([15**2]))
-    assert np.all(node_collection2.get_sizes() == np.array([20**2]))
-
-    # zorder
-    assert node_collection.get_zorder() == 0
-    assert node_collection2.get_zorder() == 10
-
-    # negative node_lw or node_size
-    with pytest.raises(ValueError):
-        ax3, node_collection3 = xgi.draw_nodes(H, node_size=-1)
-        plt.close()
-    with pytest.raises(ValueError):
-        ax3, node_collection3 = xgi.draw_nodes(H, node_lw=-1)
-        plt.close()
-
-    plt.close("all")
-
-
-def test_draw_nodes_fc_cmap(edgelist8):
-
-    H = xgi.Hypergraph(edgelist8)
-
-    # unused default when single color
-    fig, ax = plt.subplots()
-    ax, node_collection = xgi.draw_nodes(H, ax=ax, node_fc="r")
-    assert node_collection.get_cmap() == plt.cm.viridis
-    plt.close()
-
-    # default cmap
-    fig, ax = plt.subplots()
-    colors = np.linspace(11, 21, H.num_nodes)
-    ax, node_collection = xgi.draw_nodes(H, ax=ax, node_fc=colors)
-    assert node_collection.get_cmap() == plt.cm.Reds
-    plt.close()
-
-    # set cmap
-    fig, ax = plt.subplots()
-    colors = np.linspace(11, 21, H.num_nodes)
-    ax, node_collection = xgi.draw_nodes(
-        H, ax=ax, node_fc=colors, node_fc_cmap="Greens"
-    )
-    assert node_collection.get_cmap() == plt.cm.Greens
-    assert (colors.min(), colors.max()) == node_collection.get_clim()
-    plt.close()
-
-    # vmin/vmax
-    fig, ax = plt.subplots()
-    colors = np.linspace(11, 21, H.num_nodes)
[…]
