### 2023-08-18 0ecf0cf
crazy idea: reimport numpy inside the faulty tests
