### 2023-08-18 d664ca9
trying to remove dependency to np in one test
