### 2023-08-18 c5849f1
tests: new for _draw_arg_to_arr
