+++ tests/drawing/test_draw.py
@@ -4,7 +4,7 @@ import pytest
 from matplotlib import cm
 
 import xgi
-from xgi.drawing.draw import _CCW_sort, _color_arg_to_dict, _scalar_arg_to_dict
+from xgi.drawing.draw import _CCW_sort, _color_arg_to_dict, _scalar_arg_to_dict, _draw_arg_to_arr
 from xgi.exception import XGIError
 
 
@@ -22,6 +22,19 @@ def test_CCW_sort():
         == np.array([[0.037, 0.537], [0.0, 0.868], [0.791, 0.91], [0.402, 0.56]])
     )
 
+def test_draw_arg_to_arr(edgelist4):
+
+    H = xgi.Hypergraph(edgelist4)
+
+    # arg stat
+    arg = H.nodes.degree
+    degree = _draw_arg_to_arr(arg)
+    assert np.all(degree == np.array([1, 2, 3, 2, 2]))
+
+    # arg dict
+    arg_dict = {1: 1, 2: 2, 3: 3, 4: 2, 5: 2}
+    degree = _draw_arg_to_arr(arg_dict)
+    assert np.all(degree == np.array([1, 2, 3, 2, 2]))
 
 def test_scalar_arg_to_dict(edgelist4):
     ids = [1, 2, 3]
@@ -241,7 +254,7 @@ def test_draw_nodes_fc_cmap(edgelist8):
 
     # default cmap
     fig, ax = plt.subplots()
-    colors = [11, 12, 14, 16, 17, 19, 21.0]
+    colors = [11, 12, 14, 16, 17, 19, 21]
     ax, node_collection = xgi.draw_nodes(H, ax=ax, node_fc=colors)
     assert node_collection.get_cmap() == plt.cm.Reds
     plt.close()
