### 2023-08-18 d459e63
removed double import np - hopefully not necessary anymore
