### 2023-08-18 8589339
feat: changed from clipping to interp between min/max. added 'rescaling_sizes' argument. test: added for new function.
