### 2023-08-18 3765024
fix: consistent default vals in draw. feat: rescaling_sizes=True by default.
