### 2023-08-21 d834352
feat: return mappable to allow colorbar (#441)
* feat: return mappable to allow colorbar

* refact: simplified and improved draw_nodes by fully using plt.scatter

* docs: updated docstrings for draw_nodes. Removed node_ec_cmap

* fix: docstring

* fix: remove unused code

* feat: accept dict as input in draw

* add check for negative input values #272

* feat: added node_shape argument in draw_nodes

* fix: updated draw and other version to match new draw_nodes. fix: most tests

* fix: last test

* docs: added colorbar note

* tests: fixed and added more. fix: now plotting nodes with non-finit
[…]
