+++ xgi/drawing/draw.py
@@ -19,6 +19,8 @@ from .draw_utils import (
     _CCW_sort,
     _color_arg_to_dict,
     _draw_init,
+    _draw_arg_to_arr,
+    _interp_draw_arg,
     _scalar_arg_to_dict,
     _update_lims,
 )
@@ -48,6 +50,7 @@ def draw(
     node_ec="black",
     node_lw=1,
     node_size=15,
+    node_shape="o",
     max_order=None,
     node_labels=False,
     hyperedge_labels=False,
@@ -106,6 +109,9 @@ def draw(
         the radiuses are specified in the same order as the nodes are found in
         H.nodes. If NodeStat, use a monotonic linear interpolation defined between
         min_node_size and max_node_size. By default, 15.
+    node_shape :  string, optional
+        The shape of the node. Specification is as matplotlib.scatter
+        marker. Default is "o".
     max_order : int, optional
         Maximum of hyperedges to plot. If None (default), plots all orders.
     node_labels : bool or dict, optional
@@ -150,13 +156,14 @@ def draw(
     draw_hyperedge_labels
 
     """
+
     settings = {
-        "min_node_size": 10.0,
-        "max_node_size": 30.0,
+        "min_node_size": 5,
+        "max_node_size": 30,
         "min_dyad_lw": 2.0,
         "max_dyad_lw": 10.0,
-        "min_node_lw": 1.0,
-        "max_node_lw": 5.0,
+        "min_node_lw": 0,
+        "max_node_lw": 5,
         "node_fc_cmap": cm.Reds,
         "node_ec_cmap": cm.Greys,
         "edge_fc_cmap": cm.Blues,
@@ -202,17 +209,18 @@ def draw(
     else:
         raise XGIError("The input must be a SimplicialComplex or Hypergraph")
 
-    draw_nodes(
-        H,
-        pos,
-        ax,
-        node_fc,
-        node_ec,
-        node_lw,
-        node_size,
-        max_order,
-        settings,
-        node_labels,
+    ax, node_collection = draw_nodes(
+        H=H,
+        pos=pos,
+        ax=ax,
+        node_fc=node_fc,
+        node_ec=node_ec,
+        node_lw=node_lw,
+        node_size=node_size,
+        node_shape=node_shape,
+        zorder=max_order,
+        params=settings,
+        node_labels=node_labels,
         **kwargs,
     )
 
@@ -221,7 +229,7 @@ def draw(
 
     ax.set_aspect(aspect, "datalim")
 
-    return ax
+    return ax, node_collection
 
 
 def draw_nodes(
@@ -232,9 +240,14 @@ def draw_nodes(
     node_ec="black",
     node_lw=1,
     node_size=15,
+    node_shape="o",
+    node_fc_cmap="Reds",
+    vmin=None,
+    vmax=None,
     zorder=None,
-    settings=None,
+    params=dict(),
     node_labels=False,
+    rescale_sizes=True,
     **kwargs,
 ):
     """Draw the nodes of a hypergraph
@@ -249,42 +262,59 @@ def draw_nodes(
         If passed, this dictionary of positions node_id:(x,y) is used for placing the
         0-simplices.  If None (default), use the `barycenter_spring_layout` to compute
         the positions.
-    node_fc : str, dict, iterable, or NodeStat, optional
-        Color of the nodes.  If str, use the same color for all nodes.  If a dict, must
-        contain (node_id: color_str) pairs.  If other iterable, assume the colors are
-        specified in the same order as the nodes are found in H.nodes. If NodeStat, use
-        the colormap specified with node_fc_cmap. By default, "white".
-    node_ec : str, dict, iterable, or NodeStat, optional
[…]
