+++ tests/drawing/test_draw.py
@@ -1,4 +1,5 @@
 import matplotlib.pyplot as plt
+import numpy as np
 import pytest
 
 import xgi
@@ -8,43 +9,196 @@ from xgi.exception import XGIError
 def test_draw(edgelist8):
     H = xgi.Hypergraph(edgelist8)
 
-    ax = xgi.draw(H)
+    fig, ax = plt.subplots()
+    ax, node_collection = xgi.draw(H, ax=ax)
 
     # number of elements
     assert len(ax.lines) == len(H.edges.filterby("size", 2))  # dyads
     assert len(ax.patches) == len(H.edges.filterby("size", 2, mode="gt"))  # hyperedges
-    assert len(ax.collections[0].get_sizes()) == H.num_nodes  # nodes
+    offsets = node_collection.get_offsets()
+    assert offsets.shape[0] == H.num_nodes  # nodes
 
     # zorder
     for line in ax.lines:  # dyads
         assert line.get_zorder() == 3
     for patch, z in zip(ax.patches, [2, 2, 0, 2, 2]):  # hyperedges
         assert patch.get_zorder() == z
-    assert ax.collections[0].get_zorder() == 4  # nodes
+    assert node_collection.get_zorder() == 4  # nodes
 
     plt.close()
 
 
 def test_draw_nodes(edgelist8):
+
     H = xgi.Hypergraph(edgelist8)
 
-    ax = xgi.draw_nodes(H)
+    fig, ax = plt.subplots()
+    ax, node_collection = xgi.draw_nodes(H, ax=ax)
+    fig2, ax2 = plt.subplots()
+    ax2, node_collection2 = xgi.draw_nodes(
+        H,
+        ax=ax2,
+        node_fc="r",
+        node_ec="b",
+        node_lw=2,
+        node_size=20,
+        zorder=10,
+        node_shape="v",
+    )
 
     # number of elements
     assert len(ax.lines) == 0  # dyads
     assert len(ax.patches) == 0  # hyperedges
-    assert len(ax.collections[0].get_sizes()) == H.num_nodes  # nodes
+    offsets = node_collection.get_offsets()
+    assert offsets.shape[0] == H.num_nodes  # nodes
+
+    # node_fc
+    assert np.all(
+        node_collection.get_facecolor() == np.array([[1.0, 1.0, 1.0, 1.0]])
+    )  # white
+    assert np.all(
+        node_collection2.get_facecolor() == np.array([[1.0, 0.0, 0.0, 1.0]])
+    )  # blue
+
+    # node_ec
+    assert np.all(
+        node_collection.get_edgecolor() == np.array([[0.0, 0.0, 0.0, 1.0]])
+    )  # black
+    assert np.all(
+        node_collection2.get_edgecolor() == np.array([[0.0, 0.0, 1.0, 1.0]])
+    )  # red
+
+    # node_lw
+    assert np.all(node_collection.get_linewidth() == np.array([1]))
+    assert np.all(node_collection2.get_linewidth() == np.array([2]))
+
+    # node_size
+    assert np.all(node_collection.get_sizes() == np.array([15**2]))
+    assert np.all(node_collection2.get_sizes() == np.array([20**2]))
 
     # zorder
-    assert ax.collections[0].get_zorder() == 0  # nodes
+    assert node_collection.get_zorder() == 0
+    assert node_collection2.get_zorder() == 10
 
+    # negative node_lw or node_size
+    with pytest.raises(ValueError):
+        ax3, node_collection3 = xgi.draw_nodes(H, node_size=-1)
+        plt.close()
[…]

[…]
