+++ tests/utils/test_utilities.py
@@ -1,5 +1,6 @@
 import networkx as nx
 import numpy as np
+import pandas as pd
 import pytest
 from numpy import infty
 
@@ -277,34 +278,34 @@ def test_convert_labels_to_integers(
     assert DH1.edges[0]["label"] == "e1"
 
 
-def test_dist():
+def test_hist():
     vals = [1, 10, 100, 1000]
-    data = xgi.dist(vals)
-    assert len(data) == 2
-    x, y = xgi.dist(vals)
-    assert len(x) == len(y) == 10
+    df = xgi.hist(vals)
+    assert isinstance(df, pd.DataFrame)
+    assert len(df.columns) == 2
+    assert len(df.index) == 10
 
-    x, y = xgi.dist(vals, bins=3)
-
+    df = xgi.hist(vals, bins=3)
     true_x = np.array([167.5, 500.5, 833.5])
-    assert np.allclose(x, true_x)
-
+    assert np.allclose(df["bin_center"], true_x)
     true_y = np.array([3, 0, 1])
-    assert np.allclose(y, true_y)
+    assert np.allclose(df["value"], true_y)
 
-    x, y = xgi.dist(vals, bins=3, density=True)
-    assert np.allclose(x, true_x)
+    df = xgi.hist(vals, bins=3, density=True)
+    assert np.allclose(df["bin_center"], true_x)
     true_y = np.array([0.00225225, 0.0, 0.00075075])
-    assert np.allclose(y, true_y)
+    assert np.allclose(df["value"], true_y)
 
-    x, y = xgi.dist(vals, bins=3, log_binning=True)
+    df = xgi.hist(vals, bins=3, log_binning=True)
     true_x = np.array([5.5, 55.0, 550.0])
-    assert np.allclose(x, true_x)
+    assert np.allclose(df["bin_center"], true_x)
     true_y = np.array([1, 1, 2])
-    assert np.allclose(y, true_y)
+    assert np.allclose(df["value"], true_y)
 
-    x, y = xgi.dist(vals, bins=[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 1000])
+    df = xgi.hist(vals, bins=[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 1000], bin_edges=True)
     true_x = np.array([1.5, 2.5, 3.5, 4.5, 5.5, 6.5, 7.5, 8.5, 9.5, 55.0, 550.0])
-    assert np.allclose(x, true_x)
+    assert np.allclose(df["bin_center"], true_x)
     true_y = np.array([1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 2])
-    assert np.allclose(y, true_y)
+    assert np.allclose(df["value"], true_y)
+    assert np.allclose(df["bin_lo"], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100])
+    assert np.allclose(df["bin_hi"], [2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 1000])
