### 2023-08-22 b29d8f9
refact: initial rewrite of draw_hyperedges with collections
