+++ docs/source/api/stats/xgi.stats.MultiNodeStat.rst
@@ -25,3 +25,4 @@
       ~MultiNodeStat.aslist
       ~MultiNodeStat.asnumpy
       ~MultiNodeStat.aspandas
+      ~MultiNodeStat.ashist
