### 2023-08-22 2fc3b07
attempt at handling color args
