### 2023-08-22 bf2a099
Added `ashist()` method to stats (#452)
* first commit

* updates

* removed `base` argument

* added unit tests

* style: format with isort and black

* response to review

* Update __init__.py

* response to review
