+++ xgi/utils/utilities.py
@@ -5,6 +5,8 @@ from copy import deepcopy
 from functools import lru_cache
 from itertools import chain, combinations, count
 
+import numpy as np
+import pandas as pd
 import requests
 from numpy import infty
 
@@ -20,6 +22,7 @@ __all__ = [
     "request_json_from_url_cached",
     "subfaces",
     "convert_labels_to_integers",
+    "hist",
 ]
 
 
@@ -415,3 +418,66 @@ def convert_labels_to_integers(net, label_attribute="label"):
     )
 
     return temp_net
+
+
+def hist(vals, bins=10, bin_edges=False, density=False, log_binning=False):
+    """Return the distribution of a numpy array.
+
+    Parameters
+    ----------
+    vals : Numpy array
+        The array of values
+    bins : int, list, or Numpy array
+        The number of bins or the bin edges.
+        By default, 10.
+    bin_edges : bool
+        Whether to also output the min and max of each bin,
+        by default, False.
+    density : bool
+        Whether to normalize the resulting distribution.
+        By default, False.
+    log_binning : bool
+        Whether to bin the values with log-sized bins.
+        By default, False.
+
+    Returns
+    -------
+    Pandas DataFrame
+        A two-column table with "bin_center" and "value" columns,
+        where "value" is a count or a probability. If `bin_edges`
+        is True, outputs two additional columns, `bin_lo` and `bin_hi`,
+        which outputs the left and right bin edges respectively.
+
+    Notes
+    -----
+    Originally from https://github.com/jkbren/networks-and-dataviz
+
+    """
+    # We need to define the support of our distribution
+    lower_bound = np.min(vals)
+    upper_bound = np.max(vals)
+
+    if log_binning:
+        lower_bound = np.log10(lower_bound) if lower_bound > 0 else 0.0
+        upper_bound = np.log10(upper_bound)
+
+    # And the bins
+    if isinstance(bins, int):
+        if log_binning:
+            bins = np.logspace(lower_bound, upper_bound, bins + 1, base=10)
+        else:
+            bins = np.linspace(lower_bound, upper_bound, bins + 1)
+    elif not isinstance(bins, (list, np.ndarray)):
+        raise XGIError("Bins must be an integer, a list, or a numpy array.")
+
+    # Then we can compute the histogram using numpy
+    y, __ = np.histogram(vals, bins=bins, density=density)
+    # Now, we need to compute for each y the bin centers
+    x = bins[1:] - np.diff(bins) / 2.0
+
+    if bin_edges:
+        return pd.DataFrame.from_dict(
+            {"bin_center": x, "value": y, "bin_lo": bins[:-1], "bin_hi": bins[1:]}
+        )
+    else:
+        return pd.DataFrame.from_dict({"bin_center": x, "value": y})
