+++ docs/source/api/stats/xgi.stats.MultiDiEdgeStat.rst
@@ -25,3 +25,4 @@
       ~MultiDiEdgeStat.aslist
       ~MultiDiEdgeStat.asnumpy
       ~MultiDiEdgeStat.aspandas
+      ~MultiDiEdgeStat.ashist
