+++ xgi/drawing/draw.py
@@ -18,8 +18,8 @@ from ..exception import XGIError
 from .draw_utils import (
     _CCW_sort,
     _color_arg_to_dict,
-    _draw_init,
     _draw_arg_to_arr,
+    _draw_init,
     _interp_draw_arg,
     _scalar_arg_to_dict,
     _update_lims,
