### 2023-08-24 xgi-org/xgi#453 @Reckony
Add badges for repo status and python version
Badges for:
A [repostatus.org](https://www.repostatus.org/) badge,
 Python versions supported,
