### 2023-08-24 bdf7a48
feature: finished functionality rewrite of drawing hyperedges. style: isort and black
