### 2023-08-24 2c05a58
tuto: new on drawing edges
