### 2023-08-25 #115 @SimonMolinsky
@Reckony Thanks for your input!

About your questions:
1 - Yes, this should be done. Could you open an issue in the package's repository and ask the authors about similar packages? Then, could you paste the issue link in this thread to let everyone know something is happening?
2 - JOSS accepted the package in the past. You can skip those checks!
