### 2023-08-25 #115 @SimonMolinsky
@nwlandry I am curious if you have seen comments from @JonnyTran - could you elaborate here?

@JonnyTran - do those comments block the package from the acceptance? Or are those propositions for future development of the package?
