### 2023-08-25 #115 @maximelucas
Thanks everyone for your time on this. I'm one of the maintainers of XGI, I'll start replying and Nicholas can complete my answers later.

@Reckony 's questions:
 1. It's a good thing to have, you're right. We already have a list of other packages at the [end of the README](https://github.com/xgi-org/xgi#other-resources). Our [JOSS paper ](https://joss.theoj.org/papers/10.21105/joss.05162) also lists these and describes the differences of XGI with them. Were you thinking of having more than this?
2. We already published in JOSS, see previous answer.

@JonnyTran's questions:
1. It's a great idea! Thanks for the links. It's actually on our to-do list. We're in the middle of [making more tutorials](https://github.com/xgi-org/xgi/issues/414), and thought we would work on integrating them into the docs once they are done.
2. That's a good idea, we'll look into that. Right now it's possible to have a `{edge_id: node_members}` and `{node_id: edge_memberships} `dictionaries with `H.edges.members(dtype=dict)` and `H.nodes.memberships()`. Is that what you were thinking, but in a pandas dataframe?
3. I never thought of that. I would've thought that users wouldn't think they could use networkx functions. Are you thinking of a place in particular in the documentation that might be confusing regarding this, or a way to make sure it's clear? In any case it's not designed so close so as to be able to use networkx functions, but a few might incidentally work, we can check this.
