+++ docs/source/api/tutorials.rst
@@ -1,0 +1,9 @@
+##################
+Tutorials
+##################
+
+.. toctree::
+   :maxdepth: 2
+   :caption: XGI in 1 minute
+
+   tutorials/XGI in 1 minute
