+++ docs/source/api/tutorials.rst
@@ -7,3 +7,15 @@ Tutorials
    :caption: XGI in 1 minute
 
    tutorials/XGI in 1 minute
+
+.. toctree::
+   :maxdepth: 2
+   :caption: XGI in 5 minutes
+
+   tutorials/XGI in 5 minutes
+
+.. toctree::
+   :maxdepth: 2
+   :caption: XGI in 15 minutes
+
+   tutorials/XGI in 15 minutes
