### 2023-08-25 093337b
update test.yml with new location of the tutorials
