+++ docs/source/api/tutorials.rst
@@ -1,21 +1,0 @@
-##################
-Tutorials
-##################
-
-.. toctree::
-   :maxdepth: 2
-   :caption: XGI in 1 minute
-
-   tutorials/XGI in 1 minute
-
-.. toctree::
-   :maxdepth: 2
-   :caption: XGI in 5 minutes
-
-   tutorials/XGI in 5 minutes
-
-.. toctree::
-   :maxdepth: 2
-   :caption: XGI in 15 minutes
-
-   tutorials/XGI in 15 minutes
