### 2023-08-25 #115 @Reckony
Apart from issue https://github.com/xgi-org/xgi/issues/453 review done.
