### 2023-08-25 #115 @JonnyTran
> @nwlandry I am curious if you have seen comments from @JonnyTran - could you elaborate here?
>
> @JonnyTran - do those comments block the package from the acceptance? Or are those propositions for future development of the package?

I meant to comments as suggestions for future development and none are blocking.
But for comment 2, I can add an issue for `xgi.to_bipartite_pandas_dataframe(H)`
