### 2023-08-25 86842b2
Update contributors.rst
