+++ docs/source/contributors.rst
@@ -111,6 +111,13 @@ Contributors
     </p>
 
     <p>
+        <img src="https://github.com/pietrotraversa.png"
+            width="50" height="50" hspace="5"
+            style="border-radius: 50%; vertical-align: middle;"/>
+        <a href="https://github.com/pietrotraversa">@pietrotraversa</a>
+    </p>
+
+    <p>
         <img src="https://github.com/saad1282.png"
             width="50" height="50" hspace="5"
             style="border-radius: 50%; vertical-align: middle;"/>
