+++ xgi/drawing/draw_utils.py
@@ -84,18 +84,60 @@ def _update_lims(pos, ax):
 
 
 def _parse_color_arg(colors, cmap, ids, id_kind="edges"):
+    """
+    Parse and process color arguments for plotting.
 
-    # Check if edge_color is an array of floats and map to edge_cmap.
-    # This is the only case handled differently from matplotlib
+    This function is needed to handle the input formats not naturally
+    handled by matploltib's Collections: IDStat, dict, and arrays of
+    floats. All those are converted to arrays of floats and.
+
+    Parameters:
+    -----------
+    colors : color or list of colors or array-like or dict or IDStat
+        The color(s) to use. The accepted formats are the same as
+        matplotlib's scatter, with the addition of dict and IDStat.
+        Those with colors:
+        * single color as a string
+        * single color as 3- or 4-tuple
+        * list of colors of length len(ids)
+        * dict of colors containing the `ids` as keys
+        Those with numerical values (will be mapped to colors):
+        * array of floats
+        * dict of floats containing the `ids` as keys
+        * IDStat containing the `ids` as keys
+    cmap : matplotlib colormap or None
+        The colormap to use for mapping numerical values to colors.
+    ids : array-like or None
+        The IDs of the elements being plotted.
+    id_kind : str, optional
+        The kind of element IDs, "edges" (default) or "nodes".
+
+    Returns:
+    --------
+    colors : single color or ndarray
+        Processed color values for plotting.
+    colors_are_mapped : bool
+        True if the colors are mapped and need special handling. This
+        is used in draw_hyperedges to deal with Collections.
+
+    Raises:
+    -------
+    TypeError
+        If color argument is inappropriate for the provided id_kind.
+    ValueError
+        If color argument has an invalid format or length.
+
+    Notes:
+    ------
+    This function processes the color argument to ensure compatibility with
+    PatchCollection's facecolor/array and checks for correct input format and length.
+    """
 
     if id_kind == "edges" and isinstance(colors, NodeStat):
         raise TypeError("The color argument for edges cannot be a NodeStat")
     elif id_kind == "nodes" and isinstance(colors, Edgestat):
         raise TypeError("The color argument for nodes cannot be an EdgeStat")
 
-    # When drawing hyperedges we need to separate dyads and other hyperedges.
-    # But the user can input and IDstat or dict that represent all hyperedges.
-    # So we need to filter these cases.
     xsize = len(ids)
 
     if isinstance(colors, IDStat):
