### 2023-08-25 99dbfc2
fix: tests and compatbility with draw simplices
