### 2023-08-26 #115 @JonnyTran
> @JonnyTran's questions:
> 2. That's a good idea, we'll look into that. Right now it's possible to have a `{edge_id: node_members}` and `{node_id: edge_memberships} `dictionaries with `H.edges.members(dtype=dict)` and `H.nodes.memberships()`. Is that what you were thinking, but in a pandas dataframe?

Yes, misunderstood about `xgi.to_bipartite_pandas_dataframe(H)` so please nvm about that. I meant to suggest that XGI can have functions similar to [`nx.to_pandas_edgelist()`](https://networkx.org/documentation/stable/reference/generated/networkx.convert_matrix.to_pandas_edgelist.html) to export to dataframes, which is more conducive for working with other plotting libraries.

> 3. I never thought of that. I would've thought that users wouldn't think they could use networkx functions. Are you thinking of a place in particular in the documentation that might be confusing regarding this, or a way to make sure it's clear? In any case it's not designed so close so as to be able to use networkx functions, but a few might incidentally work, we can check this.

I couldn't find any particular documentation suggest compatibility, but in using an API that's similar to NetworkX's, I think some users might try the familiar method `nx.neighbors(H, n_id)` instead of `H.nodes.neighbors(n_id)`. I think it is more helpful to make it clear that XGI is not meant to be an extension of NetworkX to avoid these trials-and-errors.
