+++ xgi/drawing/draw.py
@@ -24,7 +24,11 @@ from .draw_utils import (
     _scalar_arg_to_dict,
     _update_lims,
 )
-from .layout import _augmented_projection, barycenter_spring_layout
+from .layout import (
+    _augmented_projection,
+    barycenter_spring_layout,
+    bipartite_spring_layout,
+)
 
 __all__ = [
     "draw",
@@ -36,6 +40,7 @@ __all__ = [
     "draw_hypergraph_hull",
     "draw_multilayer",
     "draw_dihypergraph",
+    "draw_bipartite",
 ]
 
 
@@ -1765,5 +1770,276 @@ def draw_dihypergraph(
 
     # compute axis limits
     _update_lims(pos, ax)
+
+    return ax
+
+
+def draw_bipartite(
+    H,
+    node_pos=None,
+    edge_pos=None,
+    ax=None,
+    dyad_color="black",
+    dyad_lw=1,
+    node_fc="white",
+    node_ec="black",
+    node_marker="o",
+    node_lw=1,
+    node_size=10,
+    edge_marker_fc="lightblue",
+    edge_marker_ec="black",
+    edge_marker="s",
+    edge_marker_lw=1,
+    edge_marker_size=10,
+    max_order=None,
+    settings=None,
+    **kwargs,
+):
+    """Draw a hypergraph as a bipartite network.
+
+    Parameters
+    ----------
+    H : Hypergraph
+        The hypergraph to draw.
+    ax : matplotlib.pyplot.axes, optional
+        Axis to draw on. If None (default), get the current axes.
+    lines_fc : str, dict, iterable, optional
+        Color of the hyperedges (lines). If str, use the same color for all hyperedges.
+        If a dict, must contain (hyperedge_id: color_str) pairs. If other iterable,
+        assume the colors are specified in the same order as the hyperedges are found
+        in DH.edges. If None (default), use the size of the hyperedges.
+    lines_lw : int, float, dict, iterable, optional
+        Line width of the hyperedges (lines). If int or float, use the same width for
+        all hyperedges. If a dict, must contain (hyperedge_id: width) pairs. If other
+        iterable, assume the widths are specified in the same order as the hyperedges
+        are found in DH.edges. By default, 1.5.
+    line_head_width : float, optional
+        Length of arrows' heads. By default, 0.05
+    node_fc : str, dict, iterable, or NodeStat, optional
+        Color of the nodes.  If str, use the same color for all nodes.  If a dict, must
+        contain (node_id: color_str) pairs.  If other iterable, assume the colors are
+        specified in the same order as the nodes are found in H.nodes. If NodeStat, use
+        the colormap specified with node_fc_cmap. By default, "white".
+    node_ec : str, dict, iterable, or NodeStat, optional
+        Color of node borders.  If str, use the same color for all nodes.  If a dict,
+        must contain (node_id: color_str) pairs.  If other iterable, assume the colors
+        are specified in the same order as the nodes are found in H.nodes. If NodeStat,
+        use the colormap specified with node_ec_cmap. By default, "black".
+    node_lw : int, float, dict, iterable, or NodeStat, optional
+        Line width of the node borders in pixels.  If int or float, use the same width
+        for all node borders.  If a dict, must contain (node_id: width) pairs.  If
+        iterable, assume the widths are specified in the same order as the nodes are
+        found in H.nodes. If NodeStat, use a monotonic linear interpolation defined
+        between min_node_lw and max_node_lw. By default, 1.
+    node_size : int, float, dict, iterable, or NodeStat, optional
[…]
