### 2023-09-01 c5e8f0a
Fix directionality of BiGG reactions (#459)
* Fixed #458.

* Update contributors.rst

* update based on comments

* Update bigg_data.py

* Update test_bigg_data.py

* Update bigg_data.py
