+++ tests/readwrite/test_bigg_data.py
@@ -10,7 +10,7 @@ def test_load_bigg_data(capfd):
     # test loading the online data
     H1 = load_bigg_data("iAF1260", cache=False)
     assert H1.num_nodes == 1668
-    assert H1.num_edges == 2382
+    assert H1.num_edges == 2953
     assert H1["name"] == "iAF1260"
     assert H1["organism"] == "Escherichia coli str. K-12 substr. MG1655"
     assert H1.nodes["2agpg161_c"] == {
@@ -35,4 +35,13 @@ def test_411():
         H = load_bigg_data("iCN718")
     assert H["name"] == "iCN718"
     assert H.num_nodes == 888
-    assert H.num_edges == 1014
+    assert H.num_edges == 1436
+
+
+def test_458():
+    H = load_bigg_data("e_coli_core")
+    assert H.edges.dimembers("FORt") == ({"for_c"}, {"for_e"})
+    assert H.edges.dimembers("PFK") == ({"atp_c", "f6p_c"}, {"adp_c", "fdp_c", "h_c"})
+
+    assert H.edges.dimembers("CO2t") == ({"co2_e"}, {"co2_c"})
+    assert H.edges.dimembers("CO2t_reverse") == ({"co2_c"}, {"co2_e"})
