### 2023-09-09 0c16910
Complement of a hypergraph (#380)
* feat: new complement_hypergraph function

* fix: rename function and run contribution pipeline

* fix: use max_edge_order shorthand

* fix: improve structure of xgi.classes.function.complement

* fix: run commit pipeline

* Update CHANGELOG.md

* Update utilities.py

* update

* fix: remove circular import

---------

Co-authored-by: Nicholas Landry <nicholas.landry.91@gmail.com>
