+++ tests/algorithms/test_properties.py
@@ -434,82 +434,3 @@ def test_incidence_density(edgelist1):
     assert abs(dens_ignore_sing(7) - 1 / 3) < tol
     with pytest.raises(ValueError):
         xgi.incidence_density(H, max_order=8, ignore_singletons=True)
-
-
-def test_subfaces(edgelist5):
-    assert xgi.subfaces(edgelist5) == [
-        (0,),
-        (1,),
-        (2,),
-        (3,),
-        (0, 1),
-        (0, 2),
-        (0, 3),
-        (1, 2),
-        (1, 3),
-        (2, 3),
-        (0, 1, 2),
-        (0, 1, 3),
-        (0, 2, 3),
-        (1, 2, 3),
-        (5,),
-        (6,),
-        (8,),
-        (6,),
-        (7,),
-        (8, 6),
-        (8, 7),
-        (6, 7),
-    ]
-
-    assert xgi.subfaces(edgelist5, order=-1) == [
-        (0, 1, 2),
-        (0, 1, 3),
-        (0, 2, 3),
-        (1, 2, 3),
-        (5,),
-        (6,),
-        (8, 6),
-        (8, 7),
-        (6, 7),
-    ]
-
-    assert xgi.subfaces(edgelist5, order=0) == [
-        (0,),
-        (1,),
-        (2,),
-        (3,),
-        (5,),
-        (6,),
-        (8,),
-        (6,),
-        (7,),
-    ]
-
-    assert xgi.subfaces(edgelist5, order=1) == [
-        (0, 1),
-        (0, 2),
-        (0, 3),
-        (1, 2),
-        (1, 3),
-        (2, 3),
-        (5, 6),
-        (8, 6),
-        (8, 7),
-        (6, 7),
-    ]
-
-    assert xgi.subfaces(edgelist5, order=2) == [
-        (0, 1, 2),
-        (0, 1, 3),
-        (0, 2, 3),
-        (1, 2, 3),
-        (8, 6, 7),
-    ]
-
-    assert xgi.subfaces(edgelist5, order=3) == [(0, 1, 2, 3)]
-
-    with pytest.raises(XGIError):
-        xgi.subfaces(
-            edgelist5, order=4
-        )  # order cannot be larger than maximum order in edgelist
