### 2023-09-10 #115 @SimonMolinsky
I've seen that you pushed a PR with [the requested changes in README.md (badges)](https://github.com/xgi-org/xgi/pull/470/commits/72a5da220c9797af0454223062af1fe4564aaa9c?short_path=b335630#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5). @nwlandry , @maximelucas could you let us know when the PR is accepted?
