### 2023-09-11 xgi-org/xgi#472 @nwlandry
Add tutorials and cookbook to the readthedocs.
See https://github.com/pyOpenSci/software-submission/issues/115#issuecomment-1638396858
