### 2023-09-11 #115 @nwlandry
Hi @SimonMolinsky! So sorry for my delayed reply on this. I will certainly let you know when the PR is merged. I have looked through all the comments and suggestions and in response, we are doing the following:
* Immediate changes with [PR #470](https://github.com/xgi-org/xgi/pull/470).
    * Adding badges for python versions and the project's status (thanks @Reckony!)
    * Adding a copyable citation bibtex entry for easy citation (thanks Simon!)
    * Adding a more helpful description of what XGI actually does (thanks @Reckony!)
* Long-term changes with [Issue #472](https://github.com/xgi-org/xgi/issues/472) and [PR #](https://github.com/xgi-org/xgi/pull/457) (Thanks @JonnyTran and Simon!!).
    * Integrating the tutorials into our website will be an extensive effort, but we are committed to continuing this work.

Because of scheduling constraints, I anticipate that we will merge the PR addressing the immediate changes sometime next week.
