### 2023-09-12 cbf5c3f
Revert "attempt at fix: tutorial not showing in the api"
This reverts commit f4a416b5cd6320291dd00bc497df01b4a2a082cc.
