+++ docs/source/api/tutorials/tutorials.rst
@@ -5,5 +5,5 @@ Tutorials
 .. toctree::
    :maxdepth: 2
    :caption: Quick introductions
-   
-   XGIin1minute
+
+   XGI in 1 minute
