### 2023-09-12 1c90401
Revert "Attempt: Tutorials in outside docs folder with nbsphinx-link"
This reverts commit 49b2c7c2be866be7382bf96601a849a967baf0c1.
