+++ requirements/documentation.txt
@@ -6,4 +6,3 @@ pillow>=8.2
 matplotlib>=3.3
 sphinx-design
 nbsphinx
-nbsphinx_link
