+++ docs/source/api/tutorials/tutorials.rst
@@ -1,9 +1,0 @@
-##################
-Tutorials
-##################
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Quick introductions
-
-   XGI in 1 minute
