+++ docs/source/api/tutorials.rst
@@ -3,19 +3,36 @@ Tutorials
 ##################
 
 .. toctree::
-   :maxdepth: 2
-   :caption: XGI in 1 minute
+   :maxdepth: 1
+   :caption: Getting started
 
    tutorials/XGI in 1 minute
-
-.. toctree::
-   :maxdepth: 2
-   :caption: XGI in 5 minutes
-
    tutorials/XGI in 5 minutes
+   tutorials/XGI in 15 minutes
+
 
 .. toctree::
-   :maxdepth: 2
-   :caption: XGI in 15 minutes
+   :maxdepth: 1
+   :caption: Focus tutorials
 
-   tutorials/XGI in 15 minutes
+   tutorials/Tutorial 1 - Basic Hypergraph Functionality
+   tutorials/Tutorial 2 - Read and Write
+   tutorials/Tutorial 3 - Basic simplicial complex usage
+   tutorials/Tutorial 4 - Generative_Models
+   tutorials/Tutorial 5 - Plotting
+   tutorials/Tutorial 6 - Statistics
+   tutorials/Tutorial 7 - Convex hulls hypergraph plotting
+   tutorials/Tutorial 8 - Directed Hypergraphs
+
+.. toctree::
+   :maxdepth: 1
+   :caption: In depth tutorials
+
+   tutorials/In Depth 1 - Drawing nodes
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Case studies
+
+   tutorials/case_study_zhang2022
+   tutorials/simplicial_kuramoto_example
