+++ docs/source/api/recipes.rst
@@ -1,0 +1,5 @@
+.. toctree::
+   :maxdepth: 1
+   :caption: Recipes
+
+   recipes/recipes
