### 2023-09-12 e3411f2
Adding title to recipes.rst
