+++ docs/source/api/recipes.rst
@@ -1,3 +1,7 @@
+##################
+Recipes
+##################
+
 .. toctree::
    :maxdepth: 1
    :caption: Recipes
