+++ docs/source/index.rst
@@ -19,7 +19,7 @@
    Tutorials <api/tutorials.rst>
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 1
    :caption: Recipes
    :hidden:
    
