### 2023-09-18 06ee1d0
add citation to RTD
