### 2023-09-20 85c36bb
Response to PyOpenSci review (#470)
* Response to PyOpenSci review

* Update README.md

* added the start of a gallery

* added higher-order explainer

* Updated paper list

* Update README.md

* add citation to RTD
