+++ README.md
@@ -1,11 +1,11 @@
-### XGI ###
+<img src='logo/logo.svg' alt="XGI" width='50%'></img>
 
-<img src='logo/logo.svg' width='150px' align="right" style="float:right;margin-left:10pt"></img>
-Comple**X** **G**roup **I**nteractions (**XGI**) is a Python package for the representation, manipulation, and study of the structure, dynamics, and functions of complex systems with group (higher-order) interactions.
-
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![Supports Python versions 3.8 and above.](https://img.shields.io/badge/Python%20versions%20supported-3.8%2B-forest
+)](https://www.repostatus.org/#active)
 [![Test Status](https://github.com/xgi-org/xgi/workflows/test/badge.svg?branch=main)](https://github.com/xgi-org/xgi/actions?query=workflow%3A%22test%22)
 [![codecov](https://codecov.io/gh/xgi-org/xgi/branch/main/graph/badge.svg?token=BI6TX2WDSG)](https://codecov.io/gh/xgi-org/xgi)
-[![Good First Issue](https://img.shields.io/badge/contribute-Good%20First%20Issue-%232EBC4F)](https://github.com/xgi-org/xgi/issues?q=is%3Aopen+is%3Aissue+label%3A%22Good+First+Issue%22)
+[![Good First Issue](https://img.shields.io/badge/contribute-Good%20First%20Issue-forest)](https://github.com/xgi-org/xgi/issues?q=is%3Aopen+is%3Aissue+label%3A%22Good+First+Issue%22)
 
 * [**Source**](../../)
 * [**Bug reports**](../../issues)
@@ -17,6 +17,7 @@ Comple**X** **G**roup **I**nteractions (**XGI**) is a Python package for the rep
 Sign up for our [mailing list](http://eepurl.com/igE6ez) and follow XGI on [Twitter](https://twitter.com/xginets) or [Mastodon](https://mathstodon.xyz/@xginets)!
 
 ### Table of Contents: ###
+- [What is XGI?](#what-is-xgi)
 - [Installation](#installation)
 - [Getting Started](#getting-started)
 - [Corresponding Data](#corresponding-data)
@@ -26,6 +27,17 @@ Sign up for our [mailing list](http://eepurl.com/igE6ez) and follow XGI on [Twit
 - [Funding](#funding)
 - [Other Resources](#other-resources)
 
+## What is XGI?
+
+Comple**X** **G**roup **I**nteractions (**XGI**) is a Python package for higher-order networks (If you want more information on what higher-order networks are, see our [brief introduction](https://xgi.readthedocs.io/en/stable/higher-order.html)).
+
+**XGI is a software designed to streamline working with higher-order networks from start to finish**. XGI can
+* Create synthetic datasets from many **generative models**
+* **Read and write** higher-order datasets in a user-friendly way
+* Represent **hypergraphs, directed hypergraphs, and simplicial complexes** with efficient and flexible data structures
+* Analyze higher-order networks with **measures and algorithms**
+* **Manipulate node and edge statistics** in a flexible and customizable way.
+* Draw higher-order networks in a variety of **visually striking ways** (See our [gallery](https://xgi.readthedocs.io/en/stable/gallery.html) for several examples.)
 
 ## Installation
 XGI runs on Python 3.8 or higher.
@@ -72,14 +84,22 @@ Some concrete ways that you can get involved:
 
 
 ## How to Cite
-We acknowledge the importance of good software to support research, and we note
-that research becomes more valuable when it is communicated effectively. To
-demonstrate the value of XGI, we ask that you cite XGI in your work.
-Currently, the best way to cite XGI is to go to our
[…]
