+++ docs/source/gallery.rst
@@ -1,0 +1,17 @@
+*******
+Gallery
+*******
+
+Here are several examples of figures made using XGI:
+
+.. figure:: assets/images/XGI_2023_Fig2.png
+    :width: 400
+    :alt: Image of a hypergraph with nodes colored and sized by different properties.
+
+    A visualization of the email-enron dataset with hyperedges of sizes 2 and 3 (all isolated nodes removed). The nodes are colored by their degree and their size proportional to the Clique motif Eigenvector Centrality.
+
+.. figure:: assets/images/Simpliciality_2023_Fig3.png
+    :width: 400
+    :alt: Image of a hypergraph with nodes colored by their simpliciality measures.
+
+    The local (A) simplicial fraction, (B) edit simpliciality, and (C) face edit simpliciality of the email-enron dataset filtered to include interactions of sizes 2 and 3. The colors of the histogram bars match the node colors on their corresponding network visualization. Nodes for which the local simpliciality is undefined are colored in grey.
