### 2023-09-20 #115 @nwlandry
Hi @SimonMolinsky! PR #470 is now merged and I published a new XGI version on PyPI to reflect this. Thanks so much.
