### 2023-09-22 83e9e4a
Added `s` parameter to `neighbors` (#450)
* Added `s` parameter to `neighbors`

* Added unit tests
