### 2023-09-29 #115 @SimonMolinsky
🎉 XGI has been approved by pyOpenSci! Thank you @nwlandry for submitting XGI and many thanks to @Reckony and @JonnyTran for reviewing this package! 😸

@nwlandry there are two tasks for you to perform:

- [ ] Add the badge for pyOpenSci peer-review to the README.md of XGI. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

<IF JOSS SUBMISSION>

- [x] <Editor Task> Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉
<IF JOSS SUBMISSION/>

---

---

@JonnyTran and @Reckony this survey [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7) is designed for reviewers too, so please fill it out when you're ready!

@nwlandry I know that your package is published in JOSS, so I've checked a few boxes for you.
