### 2023-09-29 1dedef3
add pyOpenSci badge
