### 2023-09-29 #115 @nwlandry
@SimonMolinsky Thanks! I just completed the survey and added the badge. I will bring up the blog post in our all-team meeting next week.
