### 2023-10-03 #115 @SimonMolinsky
@nwlandry The one last thing is the badge:

`Add the badge for pyOpenSci peer-review to the README.md of XGI. The badge should be [![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/115).`

Let me know when it's done.

I'm not sure if you or your friends are the part of pyOpenSci Slack community at this moment, but if not, and you want to, then let me know, and I will send invitations for you :)
