### 2023-10-04 #115 @nwlandry
Hi @SimonMolinsky, thanks for the reminder! Should be done now. I would love an invitation to the Slack and I will ask the rest of the XGI team as well. Thanks @JonnyTran and @Reckony for your helpful reviews!
