### 2023-10-09 30d5c8a
reordered edges before plotting to ensure large ones are plotted first (and appear below)
