+++ xgi/drawing/draw.py
@@ -5,6 +5,7 @@ from itertools import combinations
 
 import matplotlib.pyplot as plt
 import numpy as np
+import seaborn as sb # for cmap "crest"
 from matplotlib import cm
 from matplotlib.patches import FancyArrow
 from mpl_toolkits.mplot3d.art3d import (
