### 2023-10-09 912a279
fix: added seaborn to default requirements
