+++ requirements/default.txt
@@ -4,3 +4,4 @@ pandas>=1.3
 networkx>=2.7
 requests>=2.0
 matplotlib>=3.4
+seaborn>=0.10
