### 2023-10-11 #115 @SimonMolinsky
@nwlandry I'll close the issue soon (at the end of this week) because XGI is ready to be a part of pyOpenSci. We are honored that you have submitted your work here!

About **the blog post**: we can communicate here or by email. _I will email you about joining the pyOpenSci community._

@Reckony & @JonnyTran Thank you for your support! :raised_hands:
