### 2023-10-16 27779d3
fix: updated draw_simplices to make it consistent with draw_hyperedges
