+++ xgi/drawing/draw.py
@@ -21,6 +21,7 @@ from .. import convert
 from ..algorithms import max_edge_order, unique_edge_sizes
 from ..core import DiHypergraph, Hypergraph, SimplicialComplex
 from ..exception import XGIError
+from ..utils import subfaces
 from .draw_utils import (
     _CCW_sort,
     _color_arg_to_dict,
@@ -213,20 +214,27 @@ def draw(
         max_order = max_edge_order(H)
 
     if isinstance(H, SimplicialComplex):
-        ax = draw_simplices(
+        ax, (dyad_collection, edge_collection) = draw_simplices(
             SC=H,
             pos=pos,
             ax=ax,
             dyad_color=dyad_color,
             dyad_lw=dyad_lw,
+            dyad_style=dyad_style,
+            dyad_color_cmap=dyad_color_cmap,
+            dyad_vmin=dyad_vmin,
+            dyad_vmax=dyad_vmax,
+            alpha=alpha,
             edge_fc=edge_fc,
+            edge_fc_cmap=edge_fc_cmap,
+            edge_vmin=edge_vmin,
+            edge_vmax=edge_vmax,
             max_order=max_order,
-            settings=settings,
             hyperedge_labels=hyperedge_labels,
+            rescale_sizes=rescale_sizes,
             **kwargs,
         )
-        dyad_collection = None  # for compatibility with simplices until update
-        edge_collection = None  # for compatibility with simplices until update
+    
     elif isinstance(H, Hypergraph):
 
         ax, (dyad_collection, edge_collection) = draw_hyperedges(
@@ -685,10 +693,19 @@ def draw_simplices(
     ax=None,
     dyad_color="black",
     dyad_lw=1.5,
+    dyad_style="solid",
+    dyad_color_cmap="Greys",
+    dyad_vmin=None,
+    dyad_vmax=None,
     edge_fc=None,
+    edge_fc_cmap="crest_r",
+    edge_vmin=None,
+    edge_vmax=None,
+    alpha=0.4,
     max_order=None,
-    settings=None,
+    params=dict(),
     hyperedge_labels=False,
+    rescale_sizes=True,
     **kwargs,
 ):
     """Draw maximal simplices and pairwise faces.
@@ -761,94 +778,37 @@ def draw_simplices(
     # Plot only the maximal simplices, thus let's convert the SC to H
     H_ = convert.from_max_simplices(SC)
 
+    # add the projected pairwise interactions
+    dyads = subfaces(H_.edges.members(), order=1)
+    H_.add_edges_from(dyads)
+    H_.cleanup(multiedges=False, isolates=True, connected=False, relabel=False, in_place=True, singletons=True) # remove multi-dyads
+
     if not max_order:
         max_order = max_edge_order(H_)
 
-    ax, pos = _draw_init(H_, ax, pos)
-
-    if edge_fc is None:
-        edge_fc = H_.edges.size
-
-    if settings is None:
-        settings = {
-            "min_dyad_lw": 2.0,
-            "max_dyad_lw": 10.0,
-            "edge_fc_cmap": cm.Blues,
-            "dyad_color_cmap": cm.Greys,
-        }
-
-    settings.update(kwargs)
-
-    dyad_color = _color_arg_to_dict(dyad_color, H_.edges, settings["dyad_color_cmap"])
-    dyad_lw = _scalar_arg_to_dict(
-        dyad_lw,
-        H_.edges,
-        settings["min_dyad_lw"],
-        settings["max_dyad_lw"],
+    ax, (dyad_collection, edge_collection) = draw_hyperedges(
+        H_,
+        pos=pos,
+        ax=ax,
[…]
