### 2023-10-16 0e24377
doc: updated docstrings accordingly
