+++ tests/drawing/test_draw.py
@@ -1,6 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 import pytest
+import seaborn as sb
 
 import xgi
 from xgi.exception import XGIError
@@ -29,6 +30,31 @@ def test_draw(edgelist8):
     for patch, z in zip(ax.patches, [2, 2, 0, 2, 2]):  # hyperedges
         assert patch.get_zorder() == z
     assert node_collection.get_zorder() == 4  # nodes
+
+    plt.close()
+
+    # simplicial complex 
+    S = xgi.SimplicialComplex(edgelist8)
+
+    fig, ax = plt.subplots()
+    ax, collections = xgi.draw(S, ax=ax)
+    (node_collection, dyad_collection, edge_collection) = collections
+
+    # number of elements
+    assert len(ax.lines) == 0 
+    assert len(ax.patches) == 0 
+    offsets = node_collection.get_offsets()
+    assert offsets.shape[0] == S.num_nodes  # nodes
+    assert len(ax.collections) == 3 
+
+    assert len(dyad_collection.get_paths()) == 16  # dyads
+    assert len(edge_collection.get_paths()) == 3  # other hyperedges
+
+    # zorder
+    for line in ax.lines:  # dyads
+        assert line.get_zorder() == 3
+    for patch, z in zip(ax.patches, [0, 2, 2]):  # hyperedges
+        assert patch.get_zorder() == z
 
     plt.close()
 
@@ -204,8 +230,10 @@ def test_draw_hyperedges(edgelist8):
 
     fig, ax = plt.subplots()
     ax, collections = xgi.draw_hyperedges(H, ax=ax)
-
     (dyad_collection, edge_collection) = collections
+    fig2, ax2 = plt.subplots()
+    ax2, collections2 = xgi.draw_hyperedges(H, ax=ax2, dyad_color="r",  edge_fc="r", dyad_lw=3, dyad_style="--")
+    (dyad_collection2, edge_collection2) = collections2
 
     # number of elements
     assert len(ax.lines) == 0
@@ -219,6 +247,77 @@ def test_draw_hyperedges(edgelist8):
         assert line.get_zorder() == 3
     for patch, z in zip(ax.patches, [2, 2, 0, 2, 2]):  # hyperedges
         assert patch.get_zorder() == z
+
+    # dyad_style
+    dyad_collection.get_linestyle() == [(0.0, None)]
+    dyad_collection2.get_linestyle() == [(0.0, [5.550000000000001, 2.4000000000000004])]
+
+    # dyad_fc
+    assert np.all(
+        dyad_collection.get_color() == np.array([[0, 0, 0, 1]])
+    )  # black
+    assert np.all(
+        dyad_collection2.get_color() == np.array([[1, 0, 0, 1]])
+    )  # black
+
+    # edge_fc
+    assert np.all(
+        edge_collection.get_facecolor()[:, -1] == np.array([0.4, 0.4, 0.4, 0.4, 0.4, 0.4])
+    )
+    assert np.all(
+        edge_collection2.get_facecolor() == np.array([[1., 0., 0., 0.4]])
+    )
+
+    # edge_lw
+    assert np.all(dyad_collection.get_linewidth() == np.array([1.5]))
+    assert np.all(dyad_collection2.get_linewidth() == np.array([3]))
+    assert np.all(edge_collection.get_linewidth() == np.array([1.]))
+
+    # negative node_lw or node_size
+    with pytest.raises(ValueError):
+        ax, collections = xgi.draw_hyperedges(H, ax=ax, dyad_lw=-1)
+        (dyad_collection, edge_collection) = collections
[…]
