+++ tests/drawing/test_draw.py
@@ -33,7 +33,7 @@ def test_draw(edgelist8):
 
     plt.close()
 
-    # simplicial complex 
+    # simplicial complex
     S = xgi.SimplicialComplex(edgelist8)
 
     fig, ax = plt.subplots()
@@ -41,11 +41,11 @@ def test_draw(edgelist8):
     (node_collection, dyad_collection, edge_collection) = collections
 
     # number of elements
-    assert len(ax.lines) == 0 
-    assert len(ax.patches) == 0 
+    assert len(ax.lines) == 0
+    assert len(ax.patches) == 0
     offsets = node_collection.get_offsets()
     assert offsets.shape[0] == S.num_nodes  # nodes
-    assert len(ax.collections) == 3 
+    assert len(ax.collections) == 3
 
     assert len(dyad_collection.get_paths()) == 16  # dyads
     assert len(edge_collection.get_paths()) == 3  # other hyperedges
@@ -232,7 +232,9 @@ def test_draw_hyperedges(edgelist8):
     ax, collections = xgi.draw_hyperedges(H, ax=ax)
     (dyad_collection, edge_collection) = collections
     fig2, ax2 = plt.subplots()
-    ax2, collections2 = xgi.draw_hyperedges(H, ax=ax2, dyad_color="r",  edge_fc="r", dyad_lw=3, dyad_style="--")
+    ax2, collections2 = xgi.draw_hyperedges(
+        H, ax=ax2, dyad_color="r", edge_fc="r", dyad_lw=3, dyad_style="--"
+    )
     (dyad_collection2, edge_collection2) = collections2
 
     # number of elements
@@ -253,32 +255,26 @@ def test_draw_hyperedges(edgelist8):
     dyad_collection2.get_linestyle() == [(0.0, [5.550000000000001, 2.4000000000000004])]
 
     # dyad_fc
-    assert np.all(
-        dyad_collection.get_color() == np.array([[0, 0, 0, 1]])
-    )  # black
-    assert np.all(
-        dyad_collection2.get_color() == np.array([[1, 0, 0, 1]])
-    )  # black
+    assert np.all(dyad_collection.get_color() == np.array([[0, 0, 0, 1]]))  # black
+    assert np.all(dyad_collection2.get_color() == np.array([[1, 0, 0, 1]]))  # black
 
     # edge_fc
     assert np.all(
-        edge_collection.get_facecolor()[:, -1] == np.array([0.4, 0.4, 0.4, 0.4, 0.4, 0.4])
+        edge_collection.get_facecolor()[:, -1]
+        == np.array([0.4, 0.4, 0.4, 0.4, 0.4, 0.4])
     )
-    assert np.all(
-        edge_collection2.get_facecolor() == np.array([[1., 0., 0., 0.4]])
-    )
+    assert np.all(edge_collection2.get_facecolor() == np.array([[1.0, 0.0, 0.0, 0.4]]))
 
     # edge_lw
     assert np.all(dyad_collection.get_linewidth() == np.array([1.5]))
     assert np.all(dyad_collection2.get_linewidth() == np.array([3]))
-    assert np.all(edge_collection.get_linewidth() == np.array([1.]))
+    assert np.all(edge_collection.get_linewidth() == np.array([1.0]))
 
     # negative node_lw or node_size
     with pytest.raises(ValueError):
         ax, collections = xgi.draw_hyperedges(H, ax=ax, dyad_lw=-1)
         (dyad_collection, edge_collection) = collections
         plt.close()
-
 
     plt.close("all")
 
@@ -298,7 +294,9 @@ def test_draw_hyperedges_fc_cmap(edgelist8):
     # set cmap
[…]
