+++ docs/source/api/tutorials.rst
@@ -3,36 +3,39 @@ Tutorials
 ##################
 
 .. toctree::
-   :maxdepth: 0
-   :caption: Getting started
+   :maxdepth: 2
+   :caption: Tutorials
 
-   tutorials/XGI in 1 minute
-   tutorials/XGI in 5 minutes
-   tutorials/XGI in 15 minutes
+   .. toctree::
+      :maxdepth: 1
+      :caption: Getting Started
 
+      tutorials/XGI in 1 minute
+      tutorials/XGI in 5 minutes
+      tutorials/XGI in 15 minutes
 
-.. toctree::
-   :maxdepth: 0
-   :caption: Focus tutorials
+   .. toctree::
+      :maxdepth: 1
+      :caption: Focus Tutorials
 
-   tutorials/Tutorial 1 - Basic Hypergraph Functionality
-   tutorials/Tutorial 2 - Read and Write
-   tutorials/Tutorial 3 - Basic simplicial complex usage
-   tutorials/Tutorial 4 - Generative_Models
-   tutorials/Tutorial 5 - Plotting
-   tutorials/Tutorial 6 - Statistics
-   tutorials/Tutorial 7 - Convex hulls hypergraph plotting
-   tutorials/Tutorial 8 - Directed Hypergraphs
+      tutorials/Tutorial 1 - Basic Hypergraph Functionality
+      tutorials/Tutorial 2 - Read and Write
+      tutorials/Tutorial 3 - Basic simplicial complex usage
+      tutorials/Tutorial 4 - Generative_Models
+      tutorials/Tutorial 5 - Plotting
+      tutorials/Tutorial 6 - Statistics
+      tutorials/Tutorial 7 - Convex hulls hypergraph plotting
+      tutorials/Tutorial 8 - Directed Hypergraphs
 
-.. toctree::
-   :maxdepth: 0
-   :caption: In depth tutorials
+   .. toctree::
+      :maxdepth: 1
+      :caption: In-depth Tutorials
 
-   tutorials/In Depth 1 - Drawing nodes
+      tutorials/In Depth 1 - Drawing nodes
 
-.. toctree::
-   :maxdepth: 0
-   :caption: Case studies
+   .. toctree::
+      :maxdepth: 1
+      :caption: Case Studies
 
-   tutorials/case_study_zhang2022
-   tutorials/simplicial_kuramoto_example
+      tutorials/case_study_zhang2022
+      tutorials/simplicial_kuramoto_example
