+++ docs/source/api/tutorials.rst
@@ -1,38 +1,12 @@
+##################
+Tutorials
+##################
+
 .. toctree::
    :maxdepth: 2
-   :caption: Tutorials
 
-   .. toctree::
-      :maxdepth: 2
-      :caption: Getting Started
-
-      tutorials/XGI in 1 minute
-      tutorials/XGI in 5 minutes
-      tutorials/XGI in 15 minutes
-
-   .. toctree::
-      :maxdepth: 2
-      :caption: Focus Tutorials
-
-      tutorials/Tutorial 1 - Basic Hypergraph Functionality
-      tutorials/Tutorial 2 - Read and Write
-      tutorials/Tutorial 3 - Basic simplicial complex usage
-      tutorials/Tutorial 4 - Generative_Models
-      tutorials/Tutorial 5 - Plotting
-      tutorials/Tutorial 6 - Statistics
-      tutorials/Tutorial 7 - Convex hulls hypergraph plotting
-      tutorials/Tutorial 8 - Directed Hypergraphs
-
-   .. toctree::
-      :maxdepth: 2
-      :caption: In-depth Tutorials
-
-      tutorials/In Depth 1 - Drawing nodes
-
-   .. toctree::
-      :maxdepth: 2
-      :caption: Case Studies
-
-      tutorials/case_study_zhang2022
-      tutorials/simplicial_kuramoto_example
+   tutorials/getting_started
+   tutorials/focus_tutorials
+   tutorials/in_depth
+   tutorials/case_studies
 
