+++ docs/source/api/recipes.rst
@@ -4,6 +4,5 @@ Recipes
 
 .. toctree::
    :maxdepth: 1
-   :caption: Recipes
 
    recipes/recipes
