### 2023-10-16 1b8df54
Attempt at integrating the tutorials in API (#457)
* first attempt

* add nbsphinx in the requirements

* attempt: movimg tutorials in docs/api

* adding 5 and 15 minutes

* update test.yml with new location of the tutorials

* fix: wrong path in previous commit

* minor fix

* Attempt: Tutorials in outside docs folder with nbsphinx-link

*moved tutos to original position

* minor fix

* attempt: fix wrong path in previous commit

* attempt at fix: tutorial not showing in the api

* attempt: maybe the issue is the spaces

* Revert "minor fix"

This reverts commit 295c4bf965d938edf9bf1161faa586
[…]
