+++ docs/source/api/tutorials.rst
@@ -1,0 +1,12 @@
+##################
+Tutorials
+##################
+
+.. toctree::
+   :maxdepth: 2
+
+   tutorials/getting_started
+   tutorials/focus_tutorials
+   tutorials/in_depth
+   tutorials/case_studies
+
