+++ docs/source/api/tutorials/in_depth.rst
@@ -1,0 +1,8 @@
+##################
+In depth tutorials
+##################
+
+.. toctree::
+   :maxdepth: 1
+
+   In Depth 1 - Drawing nodes
