+++ docs/source/api/tutorials/getting_started.rst
@@ -1,0 +1,10 @@
+##################
+Getting started
+##################
+
+.. toctree::
+   :maxdepth: 1
+
+   XGI in 1 minute
+   XGI in 5 minutes
+   XGI in 15 minutes
