+++ docs/source/api/tutorials/focus_tutorials.rst
@@ -1,0 +1,16 @@
+##################
+Focus tutorials
+##################
+
+
+.. toctree::
+   :maxdepth: 1
+   
+   Tutorial 1 - Basic Hypergraph Functionality
+   Tutorial 2 - Read and Write
+   Tutorial 3 - Basic simplicial complex usage
+   Tutorial 4 - Generative_Models
+   Tutorial 5 - Plotting
+   Tutorial 6 - Statistics
+   Tutorial 7 - Convex hulls hypergraph plotting
+   Tutorial 8 - Directed Hypergraphs
