+++ docs/source/api/tutorials/case_studies.rst
@@ -3,7 +3,7 @@ Case studies
 ##################
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 1
 
    case_study_zhang2022
    simplicial_kuramoto_example
