### 2023-10-16 31b0fc8
Rewrite `draw_hyperedges()` (#456)
* refact: initial rewrite of draw_hyperedges with collections

* updated parameter handling

* attempt at handling color args

* fix: alpha

* feature: finished functionality rewrite of drawing hyperedges. style: isort and black

* tuto: new on drawing edges

* updated docstrings

* fix: tests

* fix: tests and compatbility with draw simplices

* fix: notebooks and order of nodes in lshift

* added notebook on drawing edges

* fix: impose zorder edges to be below dyads

* reordered edges before plotting to ensure large ones are plotted first (and appear below
[…]
