+++ xgi/drawing/draw.py
@@ -5,9 +5,15 @@ from itertools import combinations
 
 import matplotlib.pyplot as plt
 import numpy as np
+import seaborn as sb  # for cmap "crest"
 from matplotlib import cm
 from matplotlib.patches import FancyArrow
-from mpl_toolkits.mplot3d.art3d import Line3DCollection, Poly3DCollection
+from mpl_toolkits.mplot3d.art3d import (
+    Line3DCollection,
+    LineCollection,
+    PatchCollection,
+    Poly3DCollection,
+)
 from networkx import spring_layout
 from scipy.spatial import ConvexHull
 
@@ -15,12 +21,14 @@ from .. import convert
 from ..algorithms import max_edge_order, unique_edge_sizes
 from ..core import DiHypergraph, Hypergraph, SimplicialComplex
 from ..exception import XGIError
+from ..utils import subfaces
 from .draw_utils import (
     _CCW_sort,
     _color_arg_to_dict,
     _draw_arg_to_arr,
     _draw_init,
     _interp_draw_arg,
+    _parse_color_arg,
     _scalar_arg_to_dict,
     _update_lims,
 )
@@ -43,17 +51,26 @@ def draw(
     H,
     pos=None,
     ax=None,
-    dyad_color="black",
-    dyad_lw=1.5,
-    edge_fc=None,
     node_fc="white",
     node_ec="black",
     node_lw=1,
     node_size=15,
     node_shape="o",
     max_order=None,
+    dyad_color="black",
+    dyad_lw=1.5,
+    dyad_style="solid",
+    dyad_color_cmap="Greys",
+    dyad_vmin=None,
+    dyad_vmax=None,
+    edge_fc=None,
+    edge_fc_cmap="crest_r",
+    edge_vmin=None,
+    edge_vmax=None,
+    alpha=0.4,
     node_labels=False,
     hyperedge_labels=False,
+    rescale_sizes=True,
     aspect="equal",
     **kwargs,
 ):
@@ -69,24 +86,6 @@ def draw(
         the positions.
     ax : matplotlib.pyplot.axes, optional
         Axis to draw on. If None (default), get the current axes.
-    dyad_color : str, dict, iterable, or EdgeStat, optional
-        Color of the dyadic links.  If str, use the same color for all edges. If a dict,
-        must contain (edge_id: color_str) pairs.  If iterable, assume the colors are
-        specified in the same order as the edges are found in H.edges. If EdgeStat, use
-        a colormap (specified with dyad_color_cmap) associated to it. By default,
-        "black".
-    dyad_lw : int, float, dict, iterable, or EdgeStat, optional
-        Line width of edges of order 1 (dyadic links).  If int or float, use the same
-        width for all edges.  If a dict, must contain (edge_id: width) pairs.  If
-        iterable, assume the widths are specified in the same order as the edges are
-        found in H.edges. If EdgeStat, use a monotonic linear interpolation defined
-        between min_dyad_lw and max_dyad_lw. By default, 1.5.
-    edge_fc : str, dict, iterable, or EdgeStat, optional
-        Color of the hyperedges.  If str, use the same color for all nodes.  If a dict,
-        must contain (edge_id: color_str) pairs.  If other iterable, assume the colors
-        are specified in the same order as the hyperedges are found in H.edges. If
-        EdgeStat, use the colormap specified with edge_fc_cmap. If None (default), use
-        the H.edges.size.
     node_fc : str, dict, iterable, or NodeStat, optional
         Color of the nodes.  If str, use the same color for all nodes.  If a dict, must
         contain (node_id: color_str) pairs.  If other iterable, assume the colors are
@@ -114,6 +113,44 @@ def draw(
         marker. Default is "o".
     max_order : int, optional
[…]
