### 2023-10-17 b14fd9e
clean: removed unused cmap argument
