### 2023-10-17 f178c23
fix: settings can now be changed with single values
