+++ xgi/drawing/draw.py
@@ -1074,14 +1074,17 @@ def _draw_hull(node_pos, ax, edges_ec, facecolor, alpha, zorder, radius):
     Parameters
     ----------
     node_pos : np.array
-        nx2 dimensional array containing positions of the nodes
+        Array of dimension (n, 2) containing node positions
     ax : matplotlib.pyplot.axes
+        Axis to plot on
     edges_ec : str
         Color of the border of the convex hull
     facecolor : str
         Filling color of the convex hull
     alpha : float
         Transparency of the convex hull
+    zorder : float
+        Vertical order on which to plot
     radius : float
         Radius of the convex hull in the vicinity of the nodes.
 
@@ -1433,6 +1436,13 @@ def draw_multilayer(
     -------
     ax : matplotlib Axes3DSubplot
         The subplot with the multilayer network visualization.
+
+
+    Notes
+    -----
+    The effect of the `sep` parameter is limited by the `height` of the figure.
+    If `sep` is larger than a certain value depending on `height`, no additional
+    effect will be seen.
     """
     settings = {
         "min_node_size": 10.0,
@@ -1602,13 +1612,14 @@ def draw_multilayer(
 def draw_dihypergraph(
     DH,
     ax=None,
-    lines_fc=None,
-    lines_lw=1.5,
-    line_head_width=0.05,
     node_fc="white",
     node_ec="black",
     node_lw=1,
     node_size=15,
+    node_fc_cmap="Reds",
+    lines_fc=None,
+    lines_lw=1.5,
+    line_head_width=0.05,
     edge_marker_toggle=True,
     edge_marker_fc=None,
     edge_marker_ec=None,
@@ -1619,6 +1630,7 @@ def draw_dihypergraph(
     node_labels=False,
     hyperedge_labels=False,
     settings=None,
+    rescale_sizes=True,
     **kwargs,
 ):
     """Draw a directed hypergraph
@@ -1629,18 +1641,6 @@ def draw_dihypergraph(
         The directed hypergraph to draw.
     ax : matplotlib.pyplot.axes, optional
         Axis to draw on. If None (default), get the current axes.
-    lines_fc : str, dict, iterable, optional
-        Color of the hyperedges (lines). If str, use the same color for all hyperedges.
-        If a dict, must contain (hyperedge_id: color_str) pairs. If other iterable,
-        assume the colors are specified in the same order as the hyperedges are found
-        in DH.edges. If None (default), use the size of the hyperedges.
-    lines_lw : int, float, dict, iterable, optional
-        Line width of the hyperedges (lines). If int or float, use the same width for
-        all hyperedges. If a dict, must contain (hyperedge_id: width) pairs. If other
-        iterable, assume the widths are specified in the same order as the hyperedges
-        are found in DH.edges. By default, 1.5.
-    line_head_width : float, optional
-        Length of arrows' heads. By default, 0.05
     node_fc : str, dict, iterable, or NodeStat, optional
         Color of the nodes.  If str, use the same color for all nodes.  If a dict, must
         contain (node_id: color_str) pairs.  If other iterable, assume the colors are
@@ -1663,6 +1663,21 @@ def draw_dihypergraph(
         the radiuses are specified in the same order as the nodes are found in
         H.nodes. If NodeStat, use a monotonic linear interpolation defined between
         min_node_size and max_node_size. By default, 15.
+    node_fc_cmap : colormap
+        Colormap for mapping node colors. By default, "Reds". Ignored, if `node_fc` is
+        a str (single color).
+    lines_fc : str, dict, iterable, optional
[…]
