### 2023-10-18 97013fe
add bipartite layout
