### 2023-10-19 354c967
minor changes to tutos doc
*added missing focus tutorials
*added hyperedges in depth tutorial
