+++ docs/source/api/tutorials/in_depth.rst
@@ -6,3 +6,4 @@ In depth tutorials
    :maxdepth: 1
 
    In Depth 1 - Drawing nodes
+   In Depth 2 - Drawing hyperedges
