+++ docs/source/api/tutorials/focus_tutorials.rst
@@ -1,6 +1,6 @@
 ##################
  Focus tutorials
- ##################
+##################
 
 
  .. toctree::
