### 2023-10-19 536f592
Minor changes and corrections to tutorials in the API  (#479)
* minor changes to tutos doc

*added missing focus tutorials
*added hyperedges in depth tutorial

* Minor fixes in tutos

* minor fix

* minor fixes

* recipe for colours of different hyperedges

* added missing cell in in depth 2

* corrections by max
